==> 5_correlation_analysis.Rmd <==
---
title: "correlation_analysis"
author: "Simone Stevenson"
date: "05/05/2021"
output:
  html_document: default
  pdf_document: default
---
# Setup
## TODO: Give plots captions
## TODO: Split up chunks with descriptors
## TODO: Remove any unused variables from main manuscript objects

```{r libraries, include=FALSE}


# Using R version 4.0.3

## Clear the space
rm(list = ls()) # clear memory

# Load packages ----

# TODO ----

#' TODO: Remove non-significant correlaitons (the ones whose CI pass zero) from
#' caterpillar plots
#' TODO: Rename grouping variables to categorical variables
#' TODO: Note the categorical independence matrices contain NAs because the combinations
#' are in the wrong order - not sure how to fix?

# Correlation tools
library(corrplot)
library(ape)
library(ggpubr)

# Data handling and table reshaping
library(tidyverse)
library(tidylog)
#library(reshape2)
library(devtools)
library(data.table)
library(rlist)
# library(e1071)
library(psych)
# library(arules)


# Plotting
library(ggplot2)
library(RColorBrewer)
library(plot3D)
library(plotly)
library(viridis)
library(png)
library(gridExtra)
library(GGally)
library(ggcorrplot)
library(visdat)
library(cowplot)
library(egg)
library(ggpubr)
library(ggthemes)

# Maps
library(sf)
library(leaflet)

```

```{r inputs and outputs}

# Set input and output locations ----

sig_threshold <- 0.01 # Make it 0.01 to correct for multiple comparisons
output_date <- "2023-06-02" # Current date for analysis outputs
ms_version <- "MS_v8" #Version of MS outputs are for
section <- "correlation"
input_date <- "2022-11-24" # Date of indicator data output we are using


analysis_date <- "2023-06-02"
create_new_database_version <- FALSE # Only set to true if you want to create an entirely new version from scratch
date <- "2023-06-02"
country <- NA #"Australia" # If not subsetting, set as NA, e.g. country <- NA
analysis_inputs <- "N:/LES/Burwood/Quantitative-Ecology/Indicators-Project/ecoregions/2020-08-25_indicator_output_files"
save_outputs <- "yes" #only applies to maps, other things will always save
eco_version <- "ecoregions_2017"
parent_outputs <- "N:/LES/Burwood/Quantitative-Ecology/Indicators-Project/ecoregions/"
#eco_version <- "official_teow_wwf"
indicator_columns <- c("indicator", "year", "ecoregion_id", "raw_indicator_value")
timepoints <- c("2005", "2008")
load_map <- TRUE
indicators_to_remove <- c("RLIother", "RLILU")

# Make some colour palettes
mag20 <- magma(20)
vir20 <- viridis(20)
#show_col(mag20)
mag3 <- c("#400F73FF", "#AB337CFF", "#FA815FFF")
vir3 <- viridis(3)
vir2 <- viridis(2)

# Set up some ecoregions that we know how they should behave

east_australia <- 168 # Decline over time
amazon <- 473 # Decline over time
cardamom <- 223 # In good shape
mascarene <- 20 #Had a lot of extinctions


if (!is.na(country)) {
  
  location <- tolower(country)
  
} else {
  
  location <- "global"
  
}

# Set input data directory

# Set output directory

db_version <- tail(sort(list.files(parent_outputs)), 1)
db_interim <- list.files(file.path(parent_outputs,db_version))[
  grepl("interim",list.files(file.path(parent_outputs,db_version)))]
db_outputs <- list.files(file.path(parent_outputs,db_version))[
  grepl("database",list.files(file.path(parent_outputs,db_version)))]
ind_outputs <- list.files(file.path(parent_outputs,db_version))[
  grepl("indicator",list.files(file.path(parent_outputs,db_version)))]
analysis_outputs <- list.files(file.path(parent_outputs,db_version))[
  grepl("analysis",list.files(file.path(parent_outputs,db_version)))]

interim_outputs <- file.path(parent_outputs, db_version, db_interim)
outputs <- file.path(parent_outputs, db_version, db_outputs)

if( (length(analysis_outputs)) == 0 ) {
  
  analysis_outputs <- file.path(parent_outputs, db_version, paste(date,
                                 "_analysis_output_files",sep="") )
  
  dir.create(analysis_outputs, recursive = TRUE ) # create a new directory for today's outputs
  
  
} else {
  
  analysis_outputs <- file.path(parent_outputs, db_version, analysis_outputs)
  
}

manuscript_outputs <- "N:/LES/Burwood/Quantitative-Ecology/Indicators-Project/ecoregions/2021-05-04_manuscript_outputs"

# Folder for clean data that acts as inputs to all subsequent steps

clean_data <- file.path(manuscript_outputs, input_date, paste(ms_version,
                                 "_cleaned_data",sep=""))
  
if( !dir.exists( file.path(clean_data) ) ) {
  dir.create( file.path(clean_data), recursive = TRUE )
  
}

# Folders for outputs

if( !dir.exists( file.path(manuscript_outputs) ) ) {
  dir.create( file.path(manuscript_outputs), recursive = TRUE )
  
}

todays_outputs <- file.path(manuscript_outputs, output_date)

if( !dir.exists( file.path(todays_outputs) ) ) {
  dir.create( file.path(todays_outputs), recursive = TRUE )
  
}

main_outputs <- file.path(todays_outputs, paste(ms_version,
                                 "_main_manuscript_outputs",sep=""), section)

if( !dir.exists( file.path(main_outputs) ) ) {
  dir.create( file.path(main_outputs), recursive = TRUE )
  
}

# Folder for supporting info figures

supp_outputs <- file.path(todays_outputs, paste(ms_version,
                                 "_supporting_info_outputs",sep=""), section)
  

if( !dir.exists( file.path(supp_outputs) ) ) {
  dir.create( file.path(supp_outputs), recursive = TRUE )
  
}

# Folder for exploratory specific data

data_outputs <- file.path(todays_outputs, paste(ms_version,
                                 "_data_outputs",sep=""), section)
  
if( !dir.exists( file.path(data_outputs) ) ) {
  dir.create( file.path(data_outputs), recursive = TRUE )
  
}

```

# Load functions ----

# Function to scale the values of a vector from 0 to 1

```{r functions}

# scale_to_1 <- function(vector){
#   
#   (vector-min(vector, na.rm = TRUE))/
#     (max(vector, na.rm = TRUE)-min(vector, na.rm = TRUE))
# }
# 
# add_grouping_variable <- function(variable, indicator_data, ecoregion_data) {
#   
#   out <- indicator_data %>%
#     merge(ecoregion_data[c("ecoregion_id", variable)],
#           by = "ecoregion_id") 
#   
#   return(out)
#   
# }
# 
# make_heatmap <- function(data, directory, name) {
#   
#     heatmap <- ggplot(data, aes(x = subgroup,
#                               y = combination,
#                               fill = coefficient)) +
#     geom_tile(color = "white", height = 1) +
#     geom_text(aes(label=coefficient), size = 2) +
#     scale_fill_gradient2(limits = c(-1,1), low = "#453781FF", high = "#287D8EFF") +
#     theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
#     labs(x = "Spearman's correlation coefficient",
#          y = "Pairwise indicator comparisons") +
#     guides(fill = guide_legend(title = "Correlation\ncoefficient (r)")) +
#     theme(panel.grid.major = element_blank(),
#           panel.grid.minor = element_blank(),
#           panel.background = element_rect(fill = "grey97"),
#           legend.position = "bottom",
#           axis.title.x = element_text(size=8),
#           axis.title.y = element_text(size=8),
#           axis.text.x = element_blank(),
#           axis.ticks.x = element_blank(), 
#           axis.text.y = element_text(size= 7)) +
#     facet_grid(input_relationship ~., scales = "free_y") +
#     theme(strip.text.x = element_text(size = 7, hjust = 0))
#   
#   heatmap <- heatmap +  scale_x_discrete(label = function(x) stringr::str_trunc(x, 28)) 
#   
#   return(heatmap)
#   
#  
# }

# matrix should be indicators as columns, rownames are eco_id
# group is string for group, subgroup is a string denoting the subgroup (eg mangroves)

make_subgroup_scatterplot <- function(df, group, subgroup, group_directory, lpi) {
  
  
  subgroup <- str_replace(subgroup, " - ", "_")
  subgroup <- str_replace(subgroup, " & ", "and")
  subgroup <- str_replace(subgroup, " , ", "_")
  subgroup <- str_replace(subgroup, " . ", "_")
  subgroup <- str_replace(subgroup, " ", "_")
  subgroup <- str_replace(subgroup, "/", "_")
  
  # Set rownames back to column
  # df <- tibble::rownames_to_column(df)
  # 
  # df <- df %>%
  #       dplyr::rename(ecoregion_id = rowname)
  
  ecoregion_id <- as.numeric(df$ecoregion_id)
  
  # Get all possible combinations of indicators from the df
  indicator_combos <- combn(names(df[-1]), 2)
  
  # Now loop through the combos
  i <- i + 1
  subgroup_scatterplots <- list()
  
  for (i in seq_along(1:ncol(indicator_combos))) {
    
    print(i)
    combo <- as.vector(indicator_combos[,i])
    
    print(combo)
    
    plotname <- paste(combo[1], combo[2], subgroup, sep = "_")
    
    plotname
    
    x_axis <- df[ , grepl( combo[1] , names( df ) ) ]
    range(x_axis)
    
    y_axis <- df[ , grepl( combo[2] , names( df ) ) ]
    range(y_axis)
    
    plot_df <- as.data.frame(cbind(ecoregion_id, x_axis, y_axis))
    
    scatterplot <- ggplot(plot_df, aes(x = x_axis, y = y_axis)) +
      geom_point() +
      labs(x = combo[1],
           y = combo[2]) + 
      stat_cor(method = "spearman") +
      ggtitle(paste(group, subgroup, sep = " - ")) 
      # +
      # geom_text(label = plot_df$ecoregion_id, nudge_x = 0.1)
    
    ggsave(file.path(group_directory,
                     paste(plotname, lpi,
                           "scatterplot.png",
                           sep = "_")),
           scatterplot,  device = "png")
    
    subgroup_scatterplots[[i]] <- scatterplot
    
  }
  
  subgroup_scatterplots
  
}

# data <- correlation_input_data %>% 
#         merge(ecoregion_clusters[c("ecoregion_id",
#                                    "ECO_NAME")], by = "ecoregion_id")
# 
# col2 <- "RLI_2008"
# col1 <- "BHI_plants_2005"

# plotly_scatterplot <- function(data, col1, col2) {
#   
#   xcol <- data[,col1]
#   ycol <- data[,col2]
#   labels <- data[, c("ecoregion_id", "ECO_NAME")]
# 
#   fig <- plot_ly(data = data, x = ~ xcol, y = ~ ycol , 
#                  color = ~ cluster, 
#                  text = paste(labels[,1], labels[,2], sep = " "))
#   fig <- fig %>% 
#          layout(xaxis = xcol, yaxis = ycol)
# }

```

```{r colour palettes, eval=FALSE}

library(scales)
library(viridisLite)
show_col(viridis_pal()(20))
show_col(viridis_pal(option = "plasma")(20))
show_col(viridis_pal(option = "magma")(20))
show_col(viridis_pal(option = "inferno")(20))
show_col(viridis_pal(option = "mako")(20))
show_col(viridis_pal(option = "cividis")(20))
show_col(viridis_pal(option = "rocket")(20))
show_col(viridis_pal(option = "turbo")(20))


```
# Prepare data 

```{r data prep, include = FALSE, message = FALSE, warning = FALSE}

# Get data ----

files_paths <- list.files(clean_data, full.names = TRUE)

indicators_all_timepoints <- readRDS(files_paths[str_detect(files_paths, 
                                                            "0_indicators_all_timepoints.rds")])
indicators_raw_lpi <- readRDS(files_paths[str_detect(files_paths, 
                                                     "1_indicators_cleaned_lpi.rds")])
indicators_raw_no_lpi <- readRDS(files_paths[str_detect(files_paths, 
                                                        "2_indicators_cleaned_no_lpi.rds")])
indicators_std_no_lpi <- readRDS(files_paths[str_detect(files_paths, 
                                                        "3_indicators_cleaned_standardised_no_lpi.rds")])
indicators_std_lpi <- readRDS(files_paths[str_detect(files_paths, 
                                                     "4_indicators_cleaned_standardised_lpi.rds")])
ecoregions_numeric_raw <- readRDS(files_paths[str_detect(files_paths, 
                                                         "5_ecoregions_numeric.rds")])
ecoregions_numeric_std <- readRDS(files_paths[str_detect(files_paths, 
                                                         "6_ecoregions_numeric_standardized.rds")])
ecoregions_categorical <- readRDS(files_paths[str_detect(files_paths, 
                                                         "7_ecoregions_categorical.rds")])
ecoregion_clusters <- read.csv(files_paths[str_detect(files_paths, 
                                                      "8_ecoregion_clusters.csv")])

# Shorten the biome names

ecoregions_categorical <- ecoregions_categorical %>% 
                          mutate(biome = abbreviate(biome, 15),
                                 predominant_threat_type = abbreviate(predominant_threat_type, 15))
# Get the names of indicators

indicators <- names(indicators_raw_no_lpi)[!str_detect(names(indicators_raw_no_lpi),
                                                       "ecoregion_id")]

# Get the names of factors

grouping_variables <- names(ecoregions_categorical)
grouping_variables <- grouping_variables[!grouping_variables %in% c("ecoregion_id",
                                                       "included_in_hfp")]

# Get names of numeric variables 

numeric_variables <- names(ecoregions_numeric_raw)
numeric_variables <- numeric_variables[!str_detect(numeric_variables, 
                                                   "ecoregion_id")]

# Ancillary data

# Ecoregion map ----

if(load_map == TRUE) {
  
  ecoregion_map_all <- readRDS(paste(
    file.path("N:/LES/Burwood/Quantitative-Ecology/Indicators-Project/ecoregions/inputs",
    "Ecoregions2017valid.rds")))
  
  ecoregion_map <- ecoregion_map_all %>%
    dplyr::select(ECO_ID, ECO_NAME, OBJECTID, REALM, geometry)
  
  ecoregion_map_renamed <- ecoregion_map %>% dplyr::rename(ecoregion_id = ECO_ID)
  
  rm(ecoregion_map_all, ecoregion_map)
  
}

ecoregion_geometry <- ecoregion_map_renamed %>%  
                      dplyr::select(OBJECTID, ECO_NAME, geometry)

rm(ecoregion_map_all)

threat_scheme <- read.csv(file.path("N:/LES/Burwood/Quantitative-Ecology/Indicators-Project/ecoregions/inputs/iucn_threats/iucn_threat_classification_scheme.csv"))

headline_threats <- threat_scheme %>%
  dplyr::select(headline_name) %>%
  distinct(.) %>%
  pull()

headline_threats <- c(headline_threats, "All threats")

indicator_properties <- read.csv(file.path(analysis_inputs,
                                           "indicator_properties.csv"))

#TODO: get this out of indicator_properties
indicator_relationships <- read.csv(file.path(analysis_inputs,
                                              "indicator_input_relationships.csv"))

# Rename the relationship categories and remove the old column
indicator_relationships <- indicator_relationships %>% 
                           mutate(temp = ifelse(input_relationship == "related",
                                               "divergent", "convergent")) %>% 
                           dplyr::select(-input_relationship) %>% 
                           rename(input_relationship = temp)


```

# Time point scatterplots

```{r time scatters, echo=FALSE, warning = FALSE, message = FALSE}

# * Time by time scatterplots ----

indicators_all_timepoints <- indicators_all_timepoints %>% 
                             merge(ecoregion_clusters[c("ecoregion_id", 
                                                        "cluster",
                                                        "REALM")], 
                                   by = "ecoregion_id") %>% 
                              rename(realm = REALM)
                             
# Select the correct time point columns

BHI_data <- indicators_all_timepoints %>%
  dplyr::select(all_of(c("ecoregion_id", "realm", 
                         "BHI_plants_2005", 
                         "BHI_plants_2015",
                         "cluster")))


RLI_birds_data <- indicators_all_timepoints %>%
  dplyr::select(all_of(c("ecoregion_id", "realm", 
                         "BirdRLI_2008", 
                         "BirdRLI_2016",
                         "cluster")))


extinct_data <- indicators_all_timepoints %>%
  dplyr::select(all_of(c("ecoregion_id", "realm",
                         "extinct_2008", "extinct_2016",
                         "cluster")))

threatened_data <- indicators_all_timepoints %>%
  dplyr::select(all_of(c("ecoregion_id", "realm",
                         "threatened_2008", "threatened_2016",
                         "cluster")))

LPI_data <- indicators_all_timepoints %>%
  dplyr::select(all_of(c("ecoregion_id", "realm",
                         "LPI_2005", "LPI_2015",
                         "cluster")))

time_correlation_input_list <- list(BHI_data, RLI_birds_data, extinct_data, 
                                    threatened_data, LPI_data)


time_scatterplots <- list()

for (i in seq_along(time_correlation_input_list)) {
  
  scatterplot_data <- time_correlation_input_list[[i]]
  
  labx <- names(scatterplot_data)[3]
  laby <- names(scatterplot_data)[4]
  names(scatterplot_data) <- c("ecoregion_id", "realm","varx", "vary", "cluster")
  vargroup <- "cluster"
  
  scatterplot <- ggplot(scatterplot_data,aes(x = varx, 
                                             y = vary, 
                                             color = cluster)) +
    geom_point() +
    labs(x= labx,
         y = laby) + stat_cor(method = "spearman")
  
  ggsave(file.path(supp_outputs, paste("timestep",labx, "_x_", laby,"_NL", ".png", sep = "")),
         scatterplot, device = "png")
  
  time_scatterplots[[i]] <- scatterplot
  
}

walk(time_scatterplots, print)


```
# No LPI correlations

```{r indicator comparison data, echo=FALSE, warning = FALSE, message = FALSE}

# * Prepare correlation data ----

correlation_input_data <- indicators_raw_no_lpi %>%
                          merge(ecoregion_clusters[c("ecoregion_id",
                                                     "cluster")],
                                by = "ecoregion_id") %>% 
                          merge(ecoregions_categorical,
                                by = "ecoregion_id") %>% 
                          dplyr::select(-included_in_hfp)

names(correlation_input_data)


# Look at how groups are currently classed
# lapply(correlation_input_data, class)

# Ensure ecoregion_id is numeric so it doesn't get converted

correlation_input_data$ecoregion_id <- as.numeric(correlation_input_data$ecoregion_id)

# Convert characters to factors (should only affect grouping vars)

correlation_input_data <- correlation_input_data %>% 
  mutate(across(where(is.character), as.factor))

# Correlation plots ----

# Add clusters to the grouping variables
new_grouping_variables <- c("cluster", grouping_variables)

# Tidy up the names for plotting
all_grouping_variables <- paste("all", new_grouping_variables, sep = ".")
nice_grouping_variables <- str_to_title(gsub(".", " ", new_grouping_variables, fixed=TRUE))
nice_grouping_variables <- str_replace(nice_grouping_variables, "Hfp", "HFP")
nice_grouping_variables <- str_replace(nice_grouping_variables, "Lpi", "LPI")
nice_grouping_variables <- str_replace(nice_grouping_variables, "Rli", "RLI")
nice_grouping_variables <- str_remove(nice_grouping_variables, " Factor")

# * Individual subgroup plots ----

# Loop through the data and calculate correlation coefficients for each subgroup
# of every grouping variable


groups <- list()

group_correlation_plots <- list()

group_directories <- list()

group_coefficients <- list()

group_confidence_intervals_boot <- list()

group_matrices <- list()

group_dataframes <- list()


for (i in seq_along(new_grouping_variables)) {
  
  # Format the names of the subgroups (eg mangroves, tropical forests) 
  # in the grouping variable (eg. Biome)
  
  vargroup <- new_grouping_variables[i]
  varall <- all_grouping_variables[i]
  
  print(paste( "Producing correlation plots for", vargroup, "grouping variable", sep = " "))
  
  # Create a directory for your outputs
  
  group_directory <- file.path(supp_outputs, paste(vargroup,
                                                          "correlation_plots", sep = " "))
  
  if ( !dir.exists( group_directory ) ) {
    
    dir.create( group_directory, recursive = TRUE )
    
  }
  
  # * Remove LPI ----
  # Remove LPI because it reduces the amount of data we have by about 3/4
  
  indicators <- indicators[!indicators %in% "LPI_2005"]
  
  # Get names of columns we want to subset
  
  x <- c("ecoregion_id", indicators, vargroup)
  
  # Subset correlation inputs to just the indicators, grouping variable and eco ids
  
  group_indicators <- correlation_input_data %>%
    dplyr::select(all_of(x))
  
  all <- group_indicators
  
  # Split the full dataset by the grouping variable subgroup values (eg mangroves), 
  # each into a separate matrix
  
  group_indicator_list <- split(group_indicators, group_indicators[,vargroup])
  
  # Add the full dataset to the list too (so we can compare subgroups to global)
  
  all[, ncol(all)] <- varall
  
  group_indicator_list[[varall]] <- all
  
  subgroups <- names(group_indicator_list)
  
  # Now to format the data for each sub-group, ready to calculate correlations
  
  
  subgroup_correlation_plots <- list()
  
  subgroup_coefficients <- list()
  
  subgroup_confidence_intervals <- list()
  
  subgroup_matrices <- list()
  
  subgroup_dataframes <- list()
  
  for (j in seq_along(group_indicator_list)) {
    
    group_dataframe <- group_indicator_list[[j]] %>%
      dplyr::select(-vargroup) %>%
      na.omit(.)
    
    group_matrix <- group_indicator_list[[j]] %>%
      dplyr::select(-ecoregion_id, -vargroup) %>%
      na.omit(.)
    
    # Remove the years from the column names to make it tidier
    
    names(group_matrix) <- str_remove(names(group_matrix), "_2005")
    
    names(group_matrix) <- str_remove(names(group_matrix), "_2008")
    
    names(group_dataframe) <- c("ecoregion_id",names(group_matrix))
    
    subgroup_matrices[[j]] <- group_matrix
    subgroup_dataframes[[j]] <- group_dataframe
    
    n <- nrow(group_matrix)
    
    # Create name for plot and output file
    
    name <- paste(names(group_indicator_list)[j], "n =", n, sep = " ")
    name <- str_replace(name, "&", "and")
    name <- str_replace(name, "/", "_and_")
    name <- str_replace(name, "<", " less than ")
    name <- str_replace(name, ">", " more than ")
    name <- str_replace(name, ",", "")
    name <- gsub("[()]", "", name)
    
    # Check how many rows, don't calculate correlation if too few ecoregions in group
    if (nrow(group_matrix) < 3) {
      
      print(paste("insufficient data for", names(subgroup_matrices)[j], sep = " "))
      
    } else {
      
      # Get the correlation coefficients
      
      rs <- cor(group_matrix, method = "spearman")
      
      rs_ci <- cor.ci(group_matrix, n.iter = 2000, # n iterations taken from Ruscio et al 2008
                      p = sig_threshold, # 0.01 not 0.05 bc correcting for multiple comparisons as in Sievers et al
                      method = "spearman", plot = FALSE)
      
      # saveRDS(rs_ci, file.path(group_directory, paste(name,
      #                                                 "correlation_outputs.rds",
      #                                                 sep = " ")))
      
      subgroup_confidence_intervals[[j]] <- rs_ci
      
      subgroup_coefficients[[j]] <- rs
      
      p.mat <- cor_pmat(group_matrix, sig.level = sig_threshold)
      
      # Make a correlation heatmap for the individual sub group
      
      correlation_plot <- ggcorrplot(cor(group_matrix, method = "spearman"),
                                     title = name, 
                                     p.mat = p.mat, 
                                     type = "lower", insig = "blank",
                                     outline.color = "white",
                                     colors = c("#453781FF", "white", "#287D8EFF"),
                                     lab = TRUE)
      
      subgroup_correlation_plots[[j]] <- correlation_plot
      
      ggsave(file.path(group_directory,
                       paste(name, "heatmap.png", sep = " ")),
             correlation_plot, device = "png")
      
      print(paste("correlation complete", name, sep = " "))
      
      
    }
    
    groups[[i]] <- subgroups
    
    group_correlation_plots[[i]] <- subgroup_correlation_plots
    
    group_directories[[i]] <- group_directory
    
    group_coefficients[[i]] <- subgroup_coefficients
    
    group_confidence_intervals_boot[[i]] <- subgroup_confidence_intervals
    
    group_matrices[[i]] <- subgroup_matrices
    
    group_dataframes[[i]] <- subgroup_dataframes
    
  }
}

names(group_correlation_plots) <- new_grouping_variables
names(groups) <- new_grouping_variables
names(group_matrices) <- new_grouping_variables
names(group_coefficients) <- new_grouping_variables
names(group_confidence_intervals_boot) <- new_grouping_variables
names(group_dataframes) <- new_grouping_variables

# Check nested lists are correct lengths
length(group_correlation_plots)
lapply(group_correlation_plots, length)

# * Get coefficient dataframes ----

group_correlation_dataframes <- list()

#subgroup_correlations_dataframes <- list()

for (i in seq_along(group_correlation_plots)) {
  
  # Get the plots for one grouping variable (eg Biomes)
  
  subgroup_plots <- group_correlation_plots[[i]]
  
  names(subgroup_plots) <- groups[[i]]
  
  subgroup_plots <- list.clean(subgroup_plots)
  
  subgroups <- names(subgroup_plots)
  
  subgroup_correlations_dataframes <- list()
  
  for (j in seq_along(subgroup_plots)) {
    
    # Get the plot for one sub-group of the grouping variable (eg Mangroves)
    
    correlation_plot <- subgroup_plots[[j]]
    
    subgroup <- subgroups[j]
    
    # Get the coefficients, and add some additional grouping variables 
    # for the different types of indicator combinations (eg independent or related inputs)
    
    ## Independent data
    
    correlation_df1 <- correlation_plot$data[,1:5] %>%
      merge(indicator_properties[c("indicator","inputs")], 
            by.x = "Var1", by.y = "indicator") %>%
      dplyr::rename(input1 = inputs) %>%
      merge(indicator_properties[c("indicator","inputs")], 
            by.x = "Var2", by.y = "indicator") %>%
      dplyr::rename(input2 = inputs) %>%
      mutate(subgroup = subgroup,
             combination = paste(Var1, "x", Var2, sep = " "),
             inputs = paste(input1, "x", input2, sep = " ")) %>%
      mutate(inputs = ifelse(inputs == "land use data x iucn red list",
                             "iucn red list x land use data", inputs)) %>%
      dplyr::rename(coefficient = value) %>%
      dplyr::select(-Var2) %>%
      merge(indicator_relationships, by = "combination") %>%
      dplyr::rename(ind_group = Var1) 
    
    # Related data
    
    correlation_df2 <- correlation_plot$data[,1:5] %>%
      merge(indicator_properties[c("indicator","inputs")], 
            by.x = "Var1", by.y = "indicator") %>%
      dplyr::rename(input1 = inputs) %>%
      merge(indicator_properties[c("indicator","inputs")], 
            by.x = "Var2", by.y = "indicator") %>%
      dplyr::rename(input2 = inputs) %>%
      mutate(subgroup = subgroup,
             combination = paste(Var2, "x", Var1, sep = " "),
             inputs = paste(input1, "x", input2, sep = " ")) %>%
      mutate(inputs = ifelse(inputs == "land use data x iucn red list",
                             "iucn red list x land use data", inputs)) %>%
      dplyr::rename(coefficient = value) %>%
      dplyr::select(-Var1) %>%
      merge(indicator_relationships, by = "combination") %>%
      dplyr::rename(ind_group = Var2)
    
    # Independent and dependent data combined
    correlation_df <- rbind(correlation_df1, correlation_df2)
    
    subgroup_correlations_dataframes[[j]] <- correlation_df
    
    all_combos <- correlation_df$combination
    
  }
  
  # Combine all possible subgroups and correlation coefficients for the group
  # into one dataframe
  
  subgroup_full_dataframe <- do.call(rbind, subgroup_correlations_dataframes)
  
  # write.csv(subgroup_full_dataframe, file.path(group_directories[[i]], 
  #                                              paste(names(group_correlation_plots)[[i]],
  #                                                    "coefficient_data.csv", sep = "_")))
  # saveRDS(subgroup_full_dataframe, file.path(group_directories[[i]], 
  #                                            paste(names(group_correlation_plots)[[i]],
  #                                                  "coefficient_data.rds", sep = "_")))
  
  #group_correlation_dataframes[[i]] <- subgroup_correlations_dataframes
  group_correlation_dataframes[[i]] <- subgroup_full_dataframe
  
}

length(group_correlation_dataframes)
lapply(group_correlation_dataframes, length)

names(group_correlation_dataframes) <- new_grouping_variables


## Bootstrapped confidence intervals

group_confidence_intervals_boot <- list.clean(group_confidence_intervals_boot,
                                              recursive = TRUE)

group_ci_boot <- list()

for (i in seq_along(group_confidence_intervals_boot)) {
  
  
  subgroup_cis <- group_confidence_intervals_boot[[i]]
  
  group_name <- names(group_confidence_intervals_boot)[i]
  
  subgroup_ci_boot <- list()
  
  for (j in seq_along(subgroup_cis)) {
    
    n <- nrow(group_matrices[[i]][[j]])
    
    rho <- as.data.frame(subgroup_cis[[j]][[1]]) %>% 
      tibble::rownames_to_column() %>% 
      rename(ind_1 = rowname)
    
    
    rho_long <- rho %>% 
      pivot_longer(c("BHI_plants",
                     "BIIab",
                     "BIIri",
                     "HFP",
                     "extinct",
                     "RLI", 
                     "threatened")) %>% 
      rename(ind_2 = name,
             rs = value) %>% 
      distinct(.) %>% 
      mutate(pair = paste(ind_1, "x", ind_2, sep = " ")) %>% 
      mutate(pair_flip = paste(ind_2, "x", ind_1, sep = " ")) %>% 
      mutate(group = group_name,
             subgroup = groups[[i]][[j]])
    
    pair <- c("BHI_plants x BIIab", "BHI_plants x BIIri", 
              "BHI_plants x HFP",
              "BHI_plants x extinct" ,
              "BHI_plants x RLI", 
              "BHI_plants x threatened",
              "BIIab x BIIri", 
              "BIIab x HFP",
              "BIIab x extinct",
              "BIIab x RLI" , 
              "BIIab x threatened", 
              "BIIri x HFP","BIIri x extinct", "BIIri x RLI",
              "BIIri x threatened", "HFP x extinct","HFP x RLI", 
              "HFP x threatened", "extinct x RLI", "extinct x threatened",   
              "RLI x threatened")
    
    cis <- as.data.frame(subgroup_cis[[j]][[6]]) %>% 
      tibble::rownames_to_column() %>% 
      dplyr::select(rowname, lower, upper, p) %>% 
      rename(lower_ci = lower,
             upper_ci = upper,
             pair_abbr = rowname) %>% 
      mutate(correct_pair = pair)
    
    subgroup_ci_boot[[j]] <- rho_long %>% 
      merge(cis, by.x = "pair", 
            by.y = "correct_pair") %>% 
      mutate(n = n) %>% 
      dplyr::select(group, subgroup, ind_1, ind_2, rs, lower_ci, upper_ci,
                    n, pair)
    
  }
  
  group_ci_boot[[i]] <- subgroup_ci_boot 
  
}

confidence_intervals <- list()

for (i in seq_along(group_ci_boot)) {
  
  confidence_intervals[[i]] <- do.call(rbind, group_ci_boot[[i]])
  
}


```

# No LPI Caterpillar plots

```{r caterpillar plots, echo=FALSE, warning = FALSE, message = FALSE }

# * Make independent caterpillar plots ----

caterpillar_plots <- list()

for (i in seq_along(confidence_intervals)) {
  
  y <- confidence_intervals[[i]]
  
  
  names(y) <- c("grouping var", "subgroup", "ind1", "ind2", "rs", "lower_ci",
                "upper_ci", "n", "pair")
  
  y <- y %>%
    mutate(subgroup_label = paste(subgroup, "n =", n, sep = " ")) %>%
    filter(pair == "BHI_plants x extinct"|
             pair == "BHI_plants x RLI"|
             pair == "BHI_plants x threatened"|
             pair == "BIIab x extinct"|
             pair == "BIIab x RLI"|
             pair == "BIIab x threatened"|
             pair == "BIIri x extinct"|
             pair == "BIIri x RLI"|
             pair == "BIIri x threatened"|
             pair == "HFP x extinct"|
             pair == "HFP x RLI"|
             pair == "HFP x threatened")
  
  
  y <- y %>%
    mutate(id = paste(pair, subgroup, sep = "_"))
  
  pval <- group_correlation_dataframes[[i]] %>%
    dplyr::select(combination, coefficient, pvalue, signif, subgroup) %>%
    distinct(.) %>%
    mutate(id = paste(combination, subgroup, sep = "_"))
  
  
  plotdata <- y %>%
    merge(pval[c("id","coefficient", "pvalue", "signif")], by = "id") %>%
    dplyr::select(rs, coefficient, pvalue,signif, everything()) %>%
    filter(signif != 0)
  
  catplot <-  ggplot(plotdata, aes(rs, pair)) +
    geom_point(aes(col = pair)) +
    geom_linerange(aes(xmin = lower_ci, xmax = upper_ci,
                       col = pair)) +
    facet_wrap(~ subgroup_label) +
    geom_vline(xintercept = 0, col = "red") +
    geom_vline(xintercept = 0.3, col = "black", linetype = "dotted") +
    geom_vline(xintercept = -0.3, col = "black", linetype = "dotted") +
    geom_vline(xintercept = 0.7, col = "black", linetype = "dotted") +
    geom_vline(xintercept = -0.7, col = "black", linetype = "dotted") +
    ggtitle(new_grouping_variables[[i]]) +
    theme(axis.text.y = element_blank(),
          strip.text.x = element_text(size = 5),
          axis.text.x = element_text(size = 5),
          legend.position = "none",
          legend.text = element_text(size = 5),
          axis.ticks = element_blank()) +
    xlab("Spearman's rank correlation coefficient") + 
    ylab("Ecoregion categories") + 
    labs(color ='Ecoregion categories') +
    scale_fill_viridis(discrete=TRUE) +
    scale_color_viridis(discrete=TRUE) + 
    geom_text(aes(label = pair), size= 2, vjust = 1.5) +
    xlim(-1, 1)
  
  ggsave(file.path(group_directories[[i]],
                   paste(new_grouping_variables[[i]],
                         "caterpillar_plot.png", sep = "_")),
         catplot, device = "png", width = 12, height = 17, units = "cm")
  
  caterpillar_plots[[i]] <- catplot
  
}

names(caterpillar_plots) <- new_grouping_variables

# * Make related indicator caterpillar plots ----

related_caterpillar_plots <- list()

for (i in seq_along(confidence_intervals)) {
  
  y <- confidence_intervals[[i]]
  
  names(y) <- c("grouping var", "subgroup", "ind1", "ind2", "rs", "lower_ci",
                "upper_ci", "n", "pair")
  
  y <- y %>%
    mutate(subgroup_label = paste(subgroup, "n =", n, sep = " ")) %>%
    filter(pair == "BHI_plants x BIIab"|
             pair == "BHI_plants x BIIri"|
             pair == "BHI_plants x HFP"|
             pair == "BIIab x BIIri"|
             pair == "BIIab x HFP"|
             pair == "BIIri x HFP"|
             pair == "extinct x RLI"|
             pair == "extinct x threatened"|
             pair == "RLI x threatened")
  
  y <- y %>%
    mutate(id = paste(pair, subgroup, sep = "_"))
  
  pval <- group_correlation_dataframes[[i]] %>%
    dplyr::select(combination, coefficient, pvalue, signif, subgroup) %>%
    distinct(.) %>%
    mutate(id = paste(combination, subgroup, sep = "_"))
  
  
  plotdata <- y %>%
    merge(pval[c("id","coefficient", "pvalue", "signif")], by = "id") %>%
    dplyr::select(rs, coefficient, pvalue,signif, everything()) %>%
    filter(signif != 0)
  
  catplot <-  ggplot(plotdata, aes(rs, pair)) +
    geom_point(aes(col = pair)) +
    geom_linerange(aes(xmin = lower_ci, xmax = upper_ci,
                       col = pair)) +
    facet_wrap(~ subgroup_label) +
    geom_vline(xintercept = 0, col = "red") +
    geom_vline(xintercept = 0.3, col = "black", linetype = "dotted") +
    geom_vline(xintercept = -0.3, col = "black", linetype = "dotted") +
    geom_vline(xintercept = 0.7, col = "black", linetype = "dotted") +
    geom_vline(xintercept = -0.7, col = "black", linetype = "dotted") +
    ggtitle(new_grouping_variables[[i]]) +
    theme(axis.text.y = element_blank(),
          strip.text.x = element_text(size = 5),
          axis.text.x = element_text(size = 5),
          legend.position = "none",
          legend.text = element_text(size = 5),
          axis.ticks = element_blank()) +
    xlab("Spearman's rank correlation coefficient") + 
    ylab("Ecoregion categories") + 
    labs(color ='Ecoregion categories') +
    scale_fill_viridis(discrete=TRUE) +
    scale_color_viridis(discrete=TRUE) + 
    geom_text(aes(label = pair), size= 2, vjust = 1.5) +
    xlim(-1, 1)
  
  
  ggsave(file.path(group_directories[[i]],
                   paste(new_grouping_variables[[i]],
                         "related_catplot.png", sep = "_")),
         catplot, device = "png", width = 12, height = 17, units = "cm")
  
  related_caterpillar_plots[[i]] <- catplot
  
}

names(related_caterpillar_plots) <- new_grouping_variables

# * Correlation table ----

## Make a mega table of all correlations

pairwise_correlations_cf <- do.call(rbind, group_correlation_dataframes) %>% 
  tibble::rownames_to_column()

pairwise_correlations_ci <- do.call(rbind, confidence_intervals) %>% 
  tibble::rownames_to_column()

pairwise_correlations_no_lpi <- pairwise_correlations_cf %>% 
  merge(pairwise_correlations_ci[c("group", "rs", "lower_ci",
                                   "upper_ci","pair", "subgroup")],
        by.x = c("combination","subgroup"),
        by.y = c("pair", "subgroup")) %>% 
  dplyr::select(group, subgroup, combination, coefficient, rs,
                lower_ci, upper_ci, pvalue, signif, input_relationship) %>% 
  mutate(signif_2 = ifelse((pvalue > sig_threshold | lower_ci < 0 & upper_ci > 0),
                           0, 1))

saveRDS(pairwise_correlations_no_lpi, file.path(data_outputs, "11_correlation_dataframe.rds"))
write.csv(pairwise_correlations_no_lpi, file.path(data_outputs, "11_correlation_dataframe.csv"))

```

# No LPI Scatterplots 

```{r indicator scatterplots, echo=FALSE, warning = FALSE, message = FALSE }

# * Make subgroup scatterplots ----

# Make directories for the scatterplots (this can probably go in the make plots
# loop itself now have resolved issues with naming)

scatter_directories <- list()

for (i in 1:length(names(group_matrices))) {
  
  group <- names(group_matrices)[i]
  
  print(group)
  
  scatter_directory <- file.path(supp_outputs, paste(group,
                                                            "scatterplots", sep = "_"))
  
  if ( !dir.exists( scatter_directory ) ) {
    
    dir.create( scatter_directory, recursive = FALSE )
    
  }
  
  scatter_directories[[i]] <- scatter_directory
  
}

# For each grouping variable, for each subgroup, for each indicator pair,
# make a scatterplot

group_scatterplots <- list()

for (i in seq_along(group_dataframes)){
  
  subgroup_dataframes <- group_dataframes[[i]]
  
  scatter_directory <- scatter_directories[[i]]
  
  subgroup_scatterplots <- list()
  
  for (j in seq_along(subgroup_dataframes)) {
    
    subgroup <- groups[[i]][j]
    
    subgroup
    
    subgroup <- substring(subgroup, 1, 8)
    
    # subgroup <- str_replace(subgroup, " - ", "_")
    # subgroup <- str_replace(subgroup, " & ", "_")
    # subgroup <- str_replace(subgroup, " , ", "_")
    # subgroup <- str_replace(subgroup, " . ", "_")
    # subgroup <- str_replace(subgroup, " ", "_")
    # subgroup <- str_replace(subgroup, ", ", "_")
    # subgroup <- str_replace(subgroup, " ", "_")
    # 
    # subgroup
    
    subgroup_scatterplots[[j]] <- make_subgroup_scatterplot(subgroup_dataframes[[j]], 
                                                            names(group_dataframes)[i], 
                                                            subgroup,
                                                            scatter_directory,
                                                            "no_lpi")
    
  }
  
  group_scatterplots[[i]] <- subgroup_scatterplots
  
}

```

# With LPI correlations

```{r indicator comparison data, echo=FALSE, warning = FALSE, message = FALSE}

# * Prepare correlation data ----

correlation_input_data <- indicators_raw_lpi %>%
                          merge(ecoregion_clusters[c("ecoregion_id",
                                                     "cluster")],
                                by = "ecoregion_id") %>% 
                          merge(ecoregions_categorical,
                                by = "ecoregion_id") %>% 
                          dplyr::select(-included_in_hfp)

names(correlation_input_data)


# Look at how groups are currently classed
# lapply(correlation_input_data, class)

# Ensure ecoregion_id is numeric so it doesn't get converted

correlation_input_data$ecoregion_id <- as.numeric(correlation_input_data$ecoregion_id)

# Convert characters to factors (should only affect grouping vars)

correlation_input_data <- correlation_input_data %>% 
  mutate(across(where(is.character), as.factor))

# Correlation plots ----

# Add clusters to the grouping variables
new_grouping_variables <- c("cluster", grouping_variables)

# Tidy up the names for plotting
all_grouping_variables <- paste("all", new_grouping_variables, sep = ".")
nice_grouping_variables <- str_to_title(gsub(".", " ", new_grouping_variables, fixed=TRUE))
nice_grouping_variables <- str_replace(nice_grouping_variables, "Hfp", "HFP")
nice_grouping_variables <- str_replace(nice_grouping_variables, "Lpi", "LPI")
nice_grouping_variables <- str_replace(nice_grouping_variables, "Rli", "RLI")
nice_grouping_variables <- str_remove(nice_grouping_variables, " Factor")

# * Individual subgroup plots ----

# Loop through the data and calculate correlation coefficients for each subgroup
# of every grouping variable


groups <- list()

group_correlation_plots <- list()

group_directories <- list()

group_coefficients <- list()

group_confidence_intervals_boot <- list()

group_matrices <- list()

group_dataframes <- list()


for (i in seq_along(new_grouping_variables)) {
  
  # Format the names of the subgroups (eg mangroves, tropical forests) 
  # in the grouping variable (eg. Biome)
  
  vargroup <- new_grouping_variables[i]
  varall <- all_grouping_variables[i]
  
  print(paste( "Producing correlation plots for", vargroup, "grouping variable", sep = " "))
  
  # Create a directory for your outputs
  
  group_directory <- file.path(supp_outputs, paste(vargroup,
                                                          "correlation_plots", sep = " "))
  
  if ( !dir.exists( group_directory ) ) {
    
    dir.create( group_directory, recursive = TRUE )
    
  }
  
  # * Remove LPI ----
 
  # Get names of columns we want to subset
  
  x <- c("ecoregion_id", indicators, "LPI_2005", vargroup)
  
  # Subset correlation inputs to just the indicators, grouping variable and eco ids
  
  group_indicators <- correlation_input_data %>%
    dplyr::select(all_of(x))
  
  all <- group_indicators
  
  # Split the full dataset by the grouping variable subgroup values (eg mangroves), 
  # each into a separate matrix
  
  group_indicator_list <- split(group_indicators, group_indicators[,vargroup])
  
  # Add the full dataset to the list too (so we can compare subgroups to global)
  
  all[, ncol(all)] <- varall
  
  group_indicator_list[[varall]] <- all
  
  subgroups <- names(group_indicator_list)
  
  # Now to format the data for each sub-group, ready to calculate correlations
  
  
  subgroup_correlation_plots <- list()
  
  subgroup_coefficients <- list()
  
  subgroup_confidence_intervals <- list()
  
  subgroup_matrices <- list()
  
  subgroup_dataframes <- list()
  
  for (j in seq_along(group_indicator_list)) {
    
    group_dataframe <- group_indicator_list[[j]] %>%
      dplyr::select(-vargroup) %>%
      na.omit(.)
    
    group_matrix <- group_indicator_list[[j]] %>%
      dplyr::select(-ecoregion_id, -vargroup) %>%
      na.omit(.)
    
    # Remove the years from the column names to make it tidier
    
    names(group_matrix) <- str_remove(names(group_matrix), "_2005")
    
    names(group_matrix) <- str_remove(names(group_matrix), "_2008")
    
    names(group_dataframe) <- c("ecoregion_id",names(group_matrix))
    
    subgroup_matrices[[j]] <- group_matrix
    subgroup_dataframes[[j]] <- group_dataframe
    
    n <- nrow(group_matrix)
    
    # Create name for plot and output file
    
    name <- paste(names(group_indicator_list)[j], "n =", n, sep = " ")
    name <- str_replace(name, "&", "and")
    name <- str_replace(name, "/", "_and_")
    name <- str_replace(name, "<", " less than ")
    name <- str_replace(name, ">", " more than ")
    name <- str_replace(name, ",", "")
    name <- gsub("[()]", "", name)
    
    # Check how many rows, don't calculate correlation if too few ecoregions in group
    if (nrow(group_matrix) < 3) {
      
      print(paste("insufficient data for", names(subgroup_matrices)[j], sep = " "))
      
    } else {
      
      # Get the correlation coefficients
      
      rs <- cor(group_matrix, method = "spearman")
      
      # Get the confidence intervals
      rs_ci <- cor.ci(group_matrix, n.iter = 2000, # n iterations taken from Ruscio et al 2008
                      p = sig_threshold, # 0.01 not 0.05 bc correcting for multiple comparisons as in Sievers et al,
                      method = "spearman", plot = FALSE)
      
      # saveRDS(rs_ci, file.path(group_directory, paste(name,
      #                                                 "correlation_outputs.rds",
      #                                                 sep = " ")))
      
      subgroup_confidence_intervals[[j]] <- rs_ci
      
      subgroup_coefficients[[j]] <- rs
      
      p.mat <- cor_pmat(group_matrix)
      
      # Make a correlation heatmap for the individual sub group
      
      correlation_plot <- ggcorrplot(cor(group_matrix, method = "spearman"),
                                     title = name, 
                                     p.mat = p.mat, 
                                     type = "lower", insig = "blank",
                                     outline.color = "white",
                                     colors = c("#453781FF", "white", "#287D8EFF"),
                                     lab = TRUE)
      
      subgroup_correlation_plots[[j]] <- correlation_plot
      
      ggsave(file.path(group_directory,
                       paste(name, "heatmap_lpi.png", sep = " ")),
             correlation_plot, device = "png")
      
      print(paste("correlation complete", name, sep = " "))
      
      
    }
    
    groups[[i]] <- subgroups
    
    group_correlation_plots[[i]] <- subgroup_correlation_plots
    
    group_directories[[i]] <- group_directory
    
    group_coefficients[[i]] <- subgroup_coefficients
    
    group_confidence_intervals_boot[[i]] <- subgroup_confidence_intervals
    
    group_matrices[[i]] <- subgroup_matrices
    
    group_dataframes[[i]] <- subgroup_dataframes
    
  }
}

names(group_correlation_plots) <- new_grouping_variables
names(groups) <- new_grouping_variables
names(group_matrices) <- new_grouping_variables
names(group_coefficients) <- new_grouping_variables
names(group_confidence_intervals_boot) <- new_grouping_variables
names(group_dataframes) <- new_grouping_variables

# Check nested lists are correct lengths
length(group_correlation_plots)
lapply(group_correlation_plots, length)

# * Get coefficient dataframes ----

group_correlation_dataframes <- list()

#subgroup_correlations_dataframes <- list()

for (i in seq_along(group_correlation_plots)) {
  
  # Get the plots for one grouping variable (eg Biomes)
  
  subgroup_plots <- group_correlation_plots[[i]]
  
  names(subgroup_plots) <- groups[[i]]
  
  subgroup_plots <- list.clean(subgroup_plots)
  
  subgroups <- names(subgroup_plots)
  
  subgroup_correlations_dataframes <- list()
  
  for (j in seq_along(subgroup_plots)) {
    
    # Get the plot for one sub-group of the grouping variable (eg Mangroves)
    
    correlation_plot <- subgroup_plots[[j]]
    
    subgroup <- subgroups[j]
    
    # Get the coefficients, and add some additional grouping variables 
    # for the different types of indicator combinations (eg independent or related inputs)
    
    ## Independent data
    
    correlation_df1 <- correlation_plot$data[,1:5] %>%
      merge(indicator_properties[c("indicator","inputs")], 
            by.x = "Var1", by.y = "indicator") %>%
      dplyr::rename(input1 = inputs) %>%
      merge(indicator_properties[c("indicator","inputs")], 
            by.x = "Var2", by.y = "indicator") %>%
      dplyr::rename(input2 = inputs) %>%
      mutate(subgroup = subgroup,
             combination = paste(Var1, "x", Var2, sep = " "),
             inputs = paste(input1, "x", input2, sep = " ")) %>%
      mutate(inputs = ifelse(inputs == "land use data x iucn red list",
                             "iucn red list x land use data", inputs)) %>%
      dplyr::rename(coefficient = value) %>%
      dplyr::select(-Var2) %>%
      merge(indicator_relationships, by = "combination") %>%
      dplyr::rename(ind_group = Var1) 
    
    # Related data
    
    correlation_df2 <- correlation_plot$data[,1:5] %>%
      merge(indicator_properties[c("indicator","inputs")], 
            by.x = "Var1", by.y = "indicator") %>%
      dplyr::rename(input1 = inputs) %>%
      merge(indicator_properties[c("indicator","inputs")], 
            by.x = "Var2", by.y = "indicator") %>%
      dplyr::rename(input2 = inputs) %>%
      mutate(subgroup = subgroup,
             combination = paste(Var2, "x", Var1, sep = " "),
             inputs = paste(input1, "x", input2, sep = " ")) %>%
      mutate(inputs = ifelse(inputs == "land use data x iucn red list",
                             "iucn red list x land use data", inputs)) %>%
      dplyr::rename(coefficient = value) %>%
      dplyr::select(-Var1) %>%
      merge(indicator_relationships, by = "combination") %>%
      dplyr::rename(ind_group = Var2)
    
    # Independent and dependent data combined
    correlation_df <- rbind(correlation_df1, correlation_df2)
    
    subgroup_correlations_dataframes[[j]] <- correlation_df
    
    all_combos <- correlation_df$combination
    
  }
  
  # Combine all possible subgroups and correlation coefficients for the group
  # into one dataframe
  
  subgroup_full_dataframe <- do.call(rbind, subgroup_correlations_dataframes)
  
  # write.csv(subgroup_full_dataframe, file.path(group_directories[[i]], 
  #                                              paste(names(group_correlation_plots)[[i]],
  #                                                    "coefficient_data.csv", sep = "_")))
  # saveRDS(subgroup_full_dataframe, file.path(group_directories[[i]], 
  #                                            paste(names(group_correlation_plots)[[i]],
  #                                                  "coefficient_data.rds", sep = "_")))
  
  #group_correlation_dataframes[[i]] <- subgroup_correlations_dataframes
  group_correlation_dataframes[[i]] <- subgroup_full_dataframe
  
}

length(group_correlation_dataframes)
lapply(group_correlation_dataframes, length)

names(group_correlation_dataframes) <- new_grouping_variables


## Bootstrapped confidence intervals

group_confidence_intervals_boot <- list.clean(group_confidence_intervals_boot,
                                              recursive = TRUE)

group_ci_boot <- list()

for (i in seq_along(group_confidence_intervals_boot)) {
  
  
  subgroup_cis <- group_confidence_intervals_boot[[i]]
  
  group_name <- names(group_confidence_intervals_boot)[i]
  
  subgroup_ci_boot <- list()
  
  for (j in seq_along(subgroup_cis)) {
    
    n <- nrow(group_matrices[[i]][[j]])
    
    rho <- as.data.frame(subgroup_cis[[j]][[1]]) %>% 
      tibble::rownames_to_column() %>% 
      rename(ind_1 = rowname)
    
    
    rho_long <- rho %>% 
      pivot_longer(c("BHI_plants",
                     "BIIab",
                     "BIIri",
                     "HFP",
                     "extinct",
                     "RLI", 
                     "threatened", 
                     "LPI")) %>% 
      rename(ind_2 = name,
             rs = value) %>% 
      distinct(.) %>% 
      mutate(pair = paste(ind_1, "x", ind_2, sep = " ")) %>% 
      mutate(pair_flip = paste(ind_2, "x", ind_1, sep = " ")) %>% 
      mutate(group = group_name,
             subgroup = groups[[i]][[j]])
    
    pair <- c("BHI_plants x BIIab", 
              "BHI_plants x BIIri", 
              "BHI_plants x HFP",
              "BHI_plants x extinct" ,
              "BHI_plants x RLI", 
              "BHI_plants x threatened",
              "BHI_plants x LPI",
              "BIIab x BIIri", 
              "BIIab x HFP",
              "BIIab x extinct",
              "BIIab x RLI" , 
              "BIIab x threatened", 
              "BIIab x LPI", 
              "BIIri x HFP",
              "BIIri x extinct", 
              "BIIri x RLI",
              "BIIri x threatened",
              "BIIri x LPI",
              "HFP x extinct",
              "HFP x RLI", 
              "HFP x threatened", 
              "HFP x LPI", 
              "extinct x RLI", 
              "extinct x threatened",
              "extinct x LPI",
              "RLI x threatened", 
              "RLI x LPI",
              "threatened x LPI")
    
    cis <- as.data.frame(subgroup_cis[[j]][[6]]) %>% 
      tibble::rownames_to_column() %>% 
      dplyr::select(rowname, lower, upper, p) %>% 
      rename(lower_ci = lower,
             upper_ci = upper,
             pair_abbr = rowname) %>% 
      mutate(correct_pair = pair)
    
    subgroup_ci_boot[[j]] <- rho_long %>% 
      merge(cis, by.x = "pair", 
            by.y = "correct_pair") %>% 
      mutate(n = n) %>% 
      dplyr::select(group, subgroup, ind_1, ind_2, rs, lower_ci, upper_ci,
                    n, pair)
    
  }
  
  group_ci_boot[[i]] <- subgroup_ci_boot 
  
}

confidence_intervals <- list()

for (i in seq_along(group_ci_boot)) {
  
  confidence_intervals[[i]] <- do.call(rbind, group_ci_boot[[i]])
  
}


```

# With LPI Caterpillar plots

```{r caterpillar plots, echo=FALSE, warning = FALSE, message = FALSE }

# * Make independent caterpillar plots ----

caterpillar_plots <- list()

for (i in seq_along(confidence_intervals)) {
  
  y <- confidence_intervals[[i]]
  
  
  names(y) <- c("grouping var", "subgroup", "ind1", "ind2", "rs", "lower_ci",
                "upper_ci", "n", "pair")
  
  y <- y %>%
    mutate(subgroup_label = paste(subgroup, "n =", n, sep = " ")) %>%
    filter(pair == "BHI_plants x extinct"|
             pair == "BHI_plants x RLI"|
             pair == "BHI_plants x threatened"|
             pair == "BIIab x extinct"|
             pair == "BIIab x RLI"|
             pair == "BIIab x threatened"|
             pair == "BIIri x extinct"|
             pair == "BIIri x RLI"|
             pair == "BIIri x threatened"|
             pair == "HFP x extinct"|
             pair == "HFP x RLI"|
             pair == "HFP x threatened")
  
  
  y <- y %>%
    mutate(id = paste(pair, subgroup, sep = "_"))
  
  pval <- group_correlation_dataframes[[i]] %>%
    dplyr::select(combination, coefficient, pvalue, signif, subgroup) %>%
    distinct(.) %>%
    mutate(id = paste(combination, subgroup, sep = "_"))
  
  
  plotdata <- y %>%
    merge(pval[c("id","coefficient", "pvalue", "signif")], by = "id") %>%
    dplyr::select(rs, coefficient, pvalue,signif, everything()) %>%
    filter(signif != 0)
  
  catplot <-  ggplot(plotdata, aes(rs, pair)) +
    geom_point(aes(col = pair)) +
    geom_linerange(aes(xmin = lower_ci, xmax = upper_ci,
                       col = pair)) +
    facet_wrap(~ subgroup_label) +
    geom_vline(xintercept = 0, col = "red") +
    geom_vline(xintercept = 0.3, col = "black", linetype = "dotted") +
    geom_vline(xintercept = -0.3, col = "black", linetype = "dotted") +
    geom_vline(xintercept = 0.7, col = "black", linetype = "dotted") +
    geom_vline(xintercept = -0.7, col = "black", linetype = "dotted") +
    ggtitle(new_grouping_variables[[i]]) +
    theme(axis.text.y = element_blank(),
          strip.text.x = element_text(size = 5),
          axis.text.x = element_text(size = 5),
          legend.position = "none",
          legend.text = element_text(size = 5),
          axis.ticks = element_blank()) +
    xlab("Spearman's rank correlation coefficient") + 
    ylab("Ecoregion categories") + 
    labs(color ='Ecoregion categories') +
    scale_fill_viridis(discrete=TRUE) +
    scale_color_viridis(discrete=TRUE) + 
    geom_text(aes(label = pair), size= 2, vjust = 1.5) +
    xlim(-1, 1)
  
  ggsave(file.path(group_directories[[i]],
                   paste(new_grouping_variables[[i]],
                         "caterpillar_plot_lpi.png", sep = "_")),
         catplot, device = "png", width = 12, height = 17, units = "cm")
  
  caterpillar_plots[[i]] <- catplot
  
}

names(caterpillar_plots) <- new_grouping_variables

# * Make related indicator caterpillar plots ----

related_caterpillar_plots <- list()

for (i in seq_along(confidence_intervals)) {
  
  y <- confidence_intervals[[i]]
  
  names(y) <- c("grouping var", "subgroup", "ind1", "ind2", "rs", "lower_ci",
                "upper_ci", "n", "pair")
  
  y <- y %>%
    mutate(subgroup_label = paste(subgroup, "n =", n, sep = " ")) %>%
    filter(pair == "BHI_plants x BIIab"|
             pair == "BHI_plants x BIIri"|
             pair == "BHI_plants x HFP"|
             pair == "BIIab x BIIri"|
             pair == "BIIab x HFP"|
             pair == "BIIri x HFP"|
             pair == "extinct x RLI"|
             pair == "extinct x threatened"|
             pair == "RLI x threatened")
  
  y <- y %>%
    mutate(id = paste(pair, subgroup, sep = "_"))
  
  pval <- group_correlation_dataframes[[i]] %>%
    dplyr::select(combination, coefficient, pvalue, signif, subgroup) %>%
    distinct(.) %>%
    mutate(id = paste(combination, subgroup, sep = "_"))
  
  
  plotdata <- y %>%
    merge(pval[c("id","coefficient", "pvalue", "signif")], by = "id") %>%
    dplyr::select(rs, coefficient, pvalue,signif, everything()) %>%
    filter(signif != 0)
  
  catplot <-  ggplot(plotdata, aes(rs, pair)) +
    geom_point(aes(col = pair)) +
    geom_linerange(aes(xmin = lower_ci, xmax = upper_ci,
                       col = pair)) +
    facet_wrap(~ subgroup_label) +
    geom_vline(xintercept = 0, col = "red") +
    geom_vline(xintercept = 0.3, col = "black", linetype = "dotted") +
    geom_vline(xintercept = -0.3, col = "black", linetype = "dotted") +
    geom_vline(xintercept = 0.7, col = "black", linetype = "dotted") +
    geom_vline(xintercept = -0.7, col = "black", linetype = "dotted") +
    ggtitle(new_grouping_variables[[i]]) +
    theme(axis.text.y = element_blank(),
          strip.text.x = element_text(size = 5),
          axis.text.x = element_text(size = 5),
          legend.position = "none",
          legend.text = element_text(size = 5),
          axis.ticks = element_blank()) +
    xlab("Spearman's rank correlation coefficient") + 
    ylab("Ecoregion categories") + 
    labs(color ='Ecoregion categories') +
    scale_fill_viridis(discrete=TRUE) +
    scale_color_viridis(discrete=TRUE) + 
    geom_text(aes(label = pair), size= 2, vjust = 1.5) +
    xlim(-1, 1)
  
  
  ggsave(file.path(group_directories[[i]],
                   paste(new_grouping_variables[[i]],
                         "related_caterpillar_plot_lpi.png", sep = "_")),
         catplot, device = "png", width = 12, height = 17, units = "cm")
  
  related_caterpillar_plots[[i]] <- catplot
  
}

names(related_caterpillar_plots) <- new_grouping_variables

# * Correlation table ----

## Make a mega table of all correlations

pairwise_correlations_cf <- do.call(rbind, group_correlation_dataframes) %>% 
  tibble::rownames_to_column()

pairwise_correlations_ci <- do.call(rbind, confidence_intervals) %>% 
  tibble::rownames_to_column()

pairwise_correlations_lpi <- pairwise_correlations_cf %>% 
  merge(pairwise_correlations_ci[c("group", "rs", "lower_ci",
                                   "upper_ci","pair", "subgroup")],
        by.x = c("combination","subgroup"),
        by.y = c("pair", "subgroup")) %>% 
  dplyr::select(group, subgroup, combination, coefficient, rs,
                lower_ci, upper_ci, pvalue, signif, input_relationship) %>% 
  mutate(signif_2 = ifelse((pvalue > sig_threshold | lower_ci < 0 & upper_ci > 0),
                           0, 1))



saveRDS(pairwise_correlations_lpi, file.path(data_outputs, "12_correlation_dataframe_lpi.rds"))
write.csv(pairwise_correlations_lpi, file.path(data_outputs, "12_correlation_dataframe_lpi.csv"))

```

# Global heatmap no lpi

```{r heatmaps, include = FALSE, echo=FALSE, warning = FALSE, message = FALSE}

# Pretty heatmaps but not flexible enough

names(indicators_raw_no_lpi) <- c("ecoregion_id", "BHI_plants", "BIIab",
                                  "BIIri", "HFP", "extinct",
                                  "RLI", "threatened")

cor_inputs_no_lpi <- indicators_raw_no_lpi %>%
              dplyr::select(-ecoregion_id) %>%
              as.matrix(.)

cor_outputs_no_lpi <- cor(cor_inputs_no_lpi, method = "spearman")
p_mat = cor_pmat(cor_inputs_no_lpi, sig.level = 0.01)


all_correlations <- readRDS(file.path(data_outputs, 
                                      "11_correlation_dataframe.rds"))

head(all_correlations)



correlations_edited <- all_correlations %>% 
                       filter(subgroup == "all.cluster") %>% # Get just global coeffs
                       #Switch non-significant(pval < threshold AND 99% CI do not cross 0) coefficients to 0
                       mutate(coefficient = ifelse(signif_2 == 1, coefficient, 0)) %>%
                        mutate(corr_level = ifelse(coefficient > -0.3 & # assign the rankings
                                                  coefficient < 0, "weak",
                                           ifelse(coefficient < 0.3 & # assign the rankings
                                                  coefficient > 0, "weak",
                                           ifelse(coefficient > 0.3 &
                                                  coefficient < 0.7,
                                                  "moderate",
                                           ifelse(coefficient > 0.7 & coefficient < 1|
                                                  coefficient < -0.7 & coefficient > -1,
                                                  "strong",
                                           ifelse(coefficient < -0.3 &
                                                  coefficient > -0.7,
                                                  "moderate",
                                           ifelse(coefficient == 0,
                                                  "no-relationship",
                                           ifelse(coefficient == 1|
                                                  coefficient == -1,
                                                  "self-comparison",
                                                  "NA")))))))) %>% 
                       mutate(corr_level = factor(corr_level, levels = c("strong",
                                                                         "moderate",
                                                                         "weak",
                                                                         "no-relationship",
                                                                         "self-comparison"))) %>%
                       mutate(coeff_ch = as.character(coefficient)) %>% 
                       mutate(coeff_ch = ifelse(coeff_ch == "0", "0.00",
                                              ifelse(coeff_ch == "0.8", "0.80",
                                                coeff_ch))) %>% # Make sure all coefficients display 3 digits
                       mutate(coeff_label = ifelse(input_relationship == "convergent", # identify convergent comparisons with a *
                                                   paste(coeff_ch, " (c)", sep = "", collapse = NULL),
                                                   paste(coeff_ch, " (d)", sep = "")))

data <- as.data.frame(str_split_fixed(correlations_edited$combination, " x ", 2)) 
names(data) <- c("Ind1", "Ind2")

correlations_edited <- cbind(correlations_edited, data)  

flipped_correlations <- correlations_edited %>% 
                        mutate(Ind12 = Ind1,
                               Ind21 = Ind2) %>% 
                        dplyr::select(-Ind1, -Ind2) %>% 
                        rename(Ind1 = Ind21,
                               Ind2 = Ind12)

temp_corr <- rbind(correlations_edited, flipped_correlations) %>%
            dplyr::select(Ind1, Ind2, coefficient, corr_level, coeff_label)


Ind1 <- unique(temp_corr$Ind1)
Ind2 <- Ind1
coefficient <- rep(1, times = length(Ind1))
red <- unique(temp_corr$corr_level)[3]
corr_level <- rep("self-comparison", times = length(Ind1))
coeff_label <- rep("1.00",
                   times = length(Ind1))

self_corr <- data.frame(Ind1, Ind2, coefficient, corr_level, coeff_label)

global_heatmap_no_lpi_data <- rbind(temp_corr, self_corr) %>% 
                              mutate(Ind1 = factor(Ind1, levels = c("extinct","RLI",
                                                             "threatened",
                                                             "BIIab",
                                                             "BIIri",
                                                             "BHI_plants",
                                                             "HFP"))) %>%
                              mutate(Ind2 = factor(Ind2, levels = c("extinct","RLI",
                                                             "threatened",
                                                             "BIIab",
                                                             "BIIri",
                                                             "BHI_plants",
                                                             "HFP"))) %>% 
                              mutate(pair = paste(Ind1, Ind2))


global_heatmap_no_lpi_1 <- ggcorrplot(cor_outputs_no_lpi,
                                    colors = c("#481568FF","white", "#20A386FF"),
                                    #sig.level = 0.05,
                                    #insig = "blank",
                                    #hc.order = TRUE,
                                    p.mat = p_mat,
                                    lab = TRUE,
                                    lab_size = 6,
                                    tl.cex = 16,
                                    insig = "blank",
                                    type = "lower",
                                    outline.color = "white",
                                    ggtheme = theme(panel.background = 
                                                    element_blank())) + 
                                    labs(fill='Corr. strength') 

global_heatmap_no_lpi_1

# ggsave(file.path(main_outputs, "global_heatmap_no_lpi_pretty.png"),
#          global_heatmap_no_lpi_1, device = "png")

## GGplot version - uglier but easier to edit

heatmap_pal_gradient <- scales::seq_gradient_pal("gray30", "white", "Lab")(seq(0,1,length.out=5))
heatmap_pal_vir <- c(viridis(4), "grey80")

# Square version

global_heatmap_no_lpi <- ggplot(global_heatmap_no_lpi_data, aes(x = Ind1, 
                                                                y = Ind2, 
                                                                fill = corr_level)) +
  geom_tile(color = "white", height = 1.0, alpha = 0.75) +
  geom_text(aes(label = coeff_label), size = 7) +
  scale_fill_manual(values = heatmap_pal_vir) +
  labs(x = "Indicator 1",
       y = "Indicator 2") +
    guides(fill = guide_legend(title = "Correlation\nstrength")) +
   theme(panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          panel.background = element_rect(fill = "grey97"),
          legend.position = "bottom",
         legend.title=element_text(size=20), 
         legend.text=element_text(size=18),
          axis.title.x = element_text(size=20),
          axis.title.y = element_text(size=20),
          axis.text.x = element_text(size=18),
          axis.text.y = element_text(size=18),
          axis.ticks.x = element_blank(), 
          axis.ticks.y = element_blank())


# ggsave(file.path(main_outputs, "global_heatmap_no_lpi.png"),
#        global_heatmap_no_lpi, width = 10, height = 10, units = "in",
#        device = "png")

# Triangle version

tri_data <- flipped_correlations %>% 
      dplyr::select(Ind1, Ind2, coefficient, corr_level, coeff_label) %>% 
      rbind(self_corr) %>% 
      mutate(coefficient = ifelse(coefficient == 0.00, "", coefficient)) %>% 
                        mutate(Ind1 = factor(Ind1, 
                               levels = c("BHI_plants","BIIab", "BIIri", "HFP",
                                          "extinct", "RLI", "threatened")),
                               Ind2 = factor(Ind2, 
                               levels = c("BHI_plants","BIIab", "BIIri", "HFP",
                                          "extinct", "RLI", "threatened")))
 

pal2 <- c("#357BA2FF", "#49C1ADFF", "#DEF5E5FF", "white","grey80")

global_heatmap_no_lpi_triangle <- ggplot(tri_data, aes(x = Ind1, y = Ind2, 
                                fill = corr_level)) +
  geom_tile(color = "white", height = 1.0, alpha = 0.75) +
  geom_text(aes(label = coefficient), size = 7) +
  scale_fill_manual(values = pal2) +
  labs(x = "Indicator 1",
       y = "Indicator 2") +
    guides(fill = guide_legend(title = "Correlation\nstrength",
           nrow = 2)) +
   theme(panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          panel.background = element_rect(fill = "grey97"),
          legend.position = "bottom",
         legend.title=element_text(size=20), 
         legend.text=element_text(size=20),
          axis.title.x = element_text(size=20),
          axis.title.y = element_text(size=20),
          axis.text.x = element_text(size=20, angle = 45, hjust = 1),,
          axis.text.y = element_text(size=20),
          axis.ticks.x = element_blank(), 
          axis.ticks.y = element_blank())


ggsave(file.path(main_outputs, "global_heatmap_no_lpi_triangle.png"),
       global_heatmap_no_lpi_triangle, width = 10, height = 10, units = "in",
       device = "png")
```

# Global heatmap lpi

```{r heatmaps with lpi, include = FALSE, echo=FALSE, warning = FALSE, message = FALSE}

names(indicators_raw_lpi) <- c("ecoregion_id", "BHI_plants", "BIIab",
                                  "BIIri", "HFP", "extinct", "LPI",
                                  "RLI", "threatened")
# Pretty heatmap

## All LPI ecoregions

cor_inputs_lpi <- indicators_raw_lpi %>%
              dplyr::select(-ecoregion_id) %>%
              as.matrix(.)

cor_outputs_lpi <- cor(cor_inputs_lpi, method = "spearman")
p_mat <- cor_pmat(cor_inputs_lpi, sig.level = sig_threshold)

global_heatmap_lpi <- ggcorrplot(cor_outputs_lpi,
                                  colors = c("#481568FF","white", "#20A386FF"),
                                  sig.level = sig_threshold,
                                  #insig = "blank",
                                    #hc.order = TRUE,
                                    p.mat = p_mat,
                                    lab = TRUE,
                                    lab_size = 6,
                                    tl.cex = 16,
                                    insig = "blank",
                                    type = "lower",
                                    outline.color = "white",
                                    ggtheme = theme(panel.background = 
                                                    element_blank())) + 
                                    labs(fill='Corr. strength') 

ggsave(file.path(main_outputs, "global_heatmap_lpi_pretty.png"),
         global_heatmap_lpi, device = "png")

## Only ecoregions with > 20 pops

cor_inputs_lpi_20 <- indicators_raw_lpi %>%
                  merge(ecoregions_categorical[c("ecoregion_id",
                                                 "lpi_records_factor")]) %>% 
              dplyr::select(-ecoregion_id) %>%
              filter(lpi_records_factor == "more than 20 lpi pops") %>% 
              dplyr::select(-lpi_records_factor) %>%
              as.matrix(.)

cor_outputs_lpi_20 <- cor(cor_inputs_lpi_20, method = "spearman")
p_mat <- cor_pmat(cor_inputs_lpi_20, sig.level = sig_threshold)

global_heatmap_lpi_20 <- ggcorrplot(cor_outputs_lpi_20,
                                   colors = c("#481568FF","white", "#20A386FF"),
                                    #sig.level = 0.05,
                                    #insig = "blank",
                                    #hc.order = TRUE,
                                    p.mat = p_mat,
                                    lab = TRUE,
                                    lab_size = 6,
                                    tl.cex = 16,
                                    insig = "blank",
                                    type = "lower",
                                    outline.color = "white",
                                    ggtheme = theme(panel.background = 
                                                    element_blank())) + 
                                    labs(fill='Corr. strength') 

ggsave(file.path(main_outputs, "global_heatmap_lpi_20_pretty.png"),
         global_heatmap_lpi_20, device = "png")


# Ugly but more flexible heatmap

all_correlations <- readRDS(file.path(data_outputs, 
                                      "12_correlation_dataframe_lpi.rds"))

head(all_correlations)

correlations_edited <- all_correlations %>% 
                       filter(subgroup == "all.cluster") %>% # Get just global coeffs
                         #Switch non-significant(pval < threshold AND 95% CI do not cross 0) coefficients to 0
                       mutate(coefficient = ifelse(signif_2 == 1, coefficient, 0)) %>%                        mutate(corr_level = ifelse(coefficient > -0.3 & # assign the rankings
                                                  coefficient < 0, "weak",
                                           ifelse(coefficient < 0.3 & # assign the rankings
                                                  coefficient > 0, "weak",
                                           ifelse(coefficient > 0.3 &
                                                  coefficient < 0.7,
                                                  "moderate",
                                           ifelse(coefficient > 0.7 |
                                                  coefficient < -0.7,
                                                  "strong",
                                           ifelse(coefficient < -0.3 &
                                                  coefficient > -0.7,
                                                  "moderate",
                                           ifelse(coefficient == 0,
                                                  "no-relationship",
                                                  "NA"))))))) %>% 
                       mutate(corr_level = factor(corr_level, levels = c("strong",
                                                                         "moderate",
                                                                         "weak",
                                                                         "no-relationship"))) %>% 
                       mutate(coeff_ch = as.character(coefficient)) %>% 
                       mutate(coeff_ch = ifelse(coeff_ch == "0", "0.00",
                                              ifelse(coeff_ch == "0.8", "0.80",
                                                coeff_ch))) %>% # Make sure all coefficients display 3 digits
                       mutate(coeff_label = ifelse(input_relationship == "convergent", # identify convergent comparisons with a *
                                                   paste(coeff_ch, " (c)", sep = "", collapse = NULL),
                                                   paste(coeff_ch, " (d)", sep = "")))

data <- as.data.frame(str_split_fixed(correlations_edited$combination, " x ", 2)) 
names(data) <- c("Ind1", "Ind2")

correlations_edited <- cbind(correlations_edited, data)  

flipped_correlations <- correlations_edited %>% 
                        mutate(Ind12 = Ind1,
                               Ind21 = Ind2) %>% 
                        dplyr::select(-Ind1, -Ind2) %>% 
                        rename(Ind1 = Ind21,
                               Ind2 = Ind12)

temp_corr <- rbind(correlations_edited, flipped_correlations) %>% 
            dplyr::select(Ind1, Ind2, coefficient, corr_level, coeff_label)

Ind1 <- unique(temp_corr$Ind1)
Ind2 <- Ind1
coefficient <- rep(1, times = length(Ind1))
red <- unique(temp_corr$corr_level)[3]
corr_level <- rep("self-comparison", times = length(Ind1))
coeff_label <- rep("1.00",
                   times = length(Ind1))

self_corr <- data.frame(Ind1, Ind2, coefficient, corr_level, coeff_label)

global_heatmap_lpi_data <- rbind(temp_corr, self_corr) %>% 
                              mutate(Ind1 = factor(Ind1, 
                                    levels = c("LPI",
                                              "extinct","RLI",
                                                             "threatened",
                                                             "BIIab",
                                                             "BIIri",
                                                             "BHI_plants",
                                                             "HFP"))) %>%
                              mutate(Ind2 = factor(Ind2, levels = c("LPI",
                                                                    "extinct","RLI",
                                                             "threatened",
                                                             "BIIab",
                                                             "BIIri",
                                                             "BHI_plants",
                                                             "HFP"))) 

heatmap_pal <- scales::seq_gradient_pal("gray30", "white", "Lab")(seq(0,1,length.out=4))
heatmap_pal_vir <- c(viridis(4), "grey80")

global_heatmap_lpi <- ggplot(global_heatmap_lpi_data, 
                             aes(x = Ind1, y = Ind2, 
                                 fill = corr_level)) +
  geom_tile(color = "white", height = 1.0, alpha = 0.75) +
  geom_text(aes(label = coeff_label), size = 7) +
  scale_fill_manual(values = heatmap_pal_vir) +
  labs(x = "Indicator 1",
       y = "Indicator 2") +
    guides(fill = guide_legend(title = "Correlation\nstrength")) +
   theme(panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          panel.background = element_rect(fill = "grey97"),
          legend.position = "bottom",
          legend.title=element_text(size=20), 
          legend.text=element_text(size=18),
          axis.title.x = element_text(size=18),
          axis.title.y = element_text(size=18),
          axis.text.x = element_text(size=14),
          axis.text.y = element_text(size=14),
          axis.ticks.x = element_blank(), 
          axis.ticks.y = element_blank())


ggsave(file.path(main_outputs, "global_heatmap_lpi.png"),
       global_heatmap_lpi, width = 10, height = 10, units = "in",
       device = "png")

```

# Cluster heatmaps

```{r Cluster heatmaps, include = FALSE, echo=FALSE, warning = FALSE, message = FALSE}

# Pretty

cor_inputs_no_lpi_clusters <- indicators_raw_no_lpi %>%
              merge(ecoregion_clusters[c("ecoregion_id", "cluster")]) %>% 
              filter(cluster != "Incomplete data") 

cor_inputs_cluster_list <- split(cor_inputs_no_lpi_clusters,
                                 cor_inputs_no_lpi_clusters$cluster)

pretty_cluster_heatmaps <- list()

for ( i in seq_along(cor_inputs_cluster_list)) {

  clus <- cor_inputs_cluster_list[[i]]$cluster[1] 
  
  clus
  
  inputs <- cor_inputs_cluster_list[[i]] %>% 
            dplyr::select(-ecoregion_id, - cluster) %>%
                    as.matrix(.)

cor_outputs <- cor(inputs, method = "spearman")

p_mat = cor_pmat(inputs, sig.level = sig_threshold)

if (clus == "1") {
  
pretty_cluster_heatmaps[[i]] <- ggcorrplot(cor_outputs,
                                    colors = c("#481568FF","white", "#20A386FF"),
                                    #sig.level = 0.05,
                                    #insig = "blank",
                                    #hc.order = TRUE,
                                    p.mat = p_mat,
                                    tl.cex = 30,
                                    lab = TRUE,
                                    lab_size = 10,
                                    insig = "blank",
                                    type = "lower",
                                    outline.color = "white",
                                    ggtheme = theme(panel.background = 
                                                    element_blank(),
                                                    legend.position = "none"))
} else {
  
pretty_cluster_heatmaps[[i]] <- ggcorrplot(cor_outputs,
                                    colors = c("#481568FF","white", "#20A386FF"),
                                    #sig.level = 0.05,
                                    #insig = "blank",
                                    #hc.order = TRUE,
                                    p.mat = p_mat,
                                    tl.cex = 30,
                                    lab = TRUE,
                                    lab_size = 10,
                                    insig = "blank",
                                    type = "lower",
                                    outline.color = "white",
                                    ggtheme = theme(panel.background = 
                                                    element_blank(),
                                                    legend.position = "none",
                                                    axis.text.y=element_blank(),
                                                    axis.ticks.y=element_blank()))

}
ggsave(file.path(main_outputs, paste("cluster", clus, 
                                     "global_heatmap_lpi.png", sep = "_")),
       pretty_cluster_heatmaps[[i]], width = 10, height = 10, units = "in",
       device = "png")

}

cluster_heatmaps <- plot_grid(pretty_cluster_heatmaps[[1]],
                        pretty_cluster_heatmaps[[2]],
                        pretty_cluster_heatmaps[[3]],
                        nrow = 1, align = "v")

ggsave(file.path(main_outputs, "cluster_heatmaps_pretty.png"),
       cluster_heatmaps, width = 30, height = 12, units = "in",
       device = "png")

# Ugly but flexible


all_correlations <- readRDS(file.path(data_outputs, 
                                      "11_correlation_dataframe.rds"))


correlations_clusters <- all_correlations %>% 
                       filter(group == "cluster") %>% # Get all cluster subgroup coeffs
                       filter(subgroup != "all.cluster") %>% # Get all cluster subgroup coeffs
    #Switch non-significant(pval < threshold AND 95% CI do not cross 0) coefficients to 0
                       mutate(coefficient = ifelse(signif_2 == 1, coefficient, 0)) %>%                       mutate(corr_level = ifelse(coefficient > -0.3 & # assign the rankings
                                                  coefficient < 0, "weak",
                                           ifelse(coefficient < 0.3 & # assign the rankings
                                                  coefficient > 0, "weak",
                                           ifelse(coefficient >= 0.3 &
                                                  coefficient < 0.7,
                                                  "moderate",
                                           ifelse(coefficient >= 0.7 |
                                                  coefficient <= -0.7|
                                                  coefficient == 1,
                                                  "strong",
                                           ifelse(coefficient <= -0.3 &
                                                  coefficient >= -0.7,
                                                  "moderate",
                                           ifelse(coefficient == 0,
                                                  "no-relationship",
                                                  "NA"))))))) %>% 
                       mutate(corr_level = factor(corr_level, levels = c("strong",
                                                                         "moderate",
                                                                         "weak",
                                                                         "no-relationship"))) %>% 
                       mutate(coeff_ch = as.character(coefficient)) %>% 
                       mutate(coeff_ch = ifelse(coeff_ch == "0", "0.00",
                                              ifelse(coeff_ch == "0.8", "0.80",
                                                coeff_ch))) %>% # Make sure all coefficients display 3 digits
                       mutate(coeff_label = ifelse(input_relationship == "convergent", # identify convergent comparisons with a *
                                                   paste(coeff_ch, " (c)", sep = "", collapse = NULL),
                                                   paste(coeff_ch, " (d)", sep = "")))

# Function to prep data

## Square heatmap

prep_heatmap_data <- function(data, sub_name) {
  
inds <- as.data.frame(str_split_fixed(data$combination, " x ", 2)) 

names(inds) <- c("Ind1", "Ind2")

newdata <- cbind(data, inds)  

newdata <- newdata %>% 
           mutate(subgroup = sub_name)

flipped_correlations <- newdata %>% 
                        mutate(Ind12 = Ind1,
                               Ind21 = Ind2) %>% 
                        dplyr::select(-Ind1, -Ind2) %>% 
                        rename(Ind1 = Ind21,
                               Ind2 = Ind12)

temp_corr <- rbind(newdata, flipped_correlations) %>% 
            dplyr::select(subgroup, Ind1, Ind2, coefficient, corr_level, coeff_label)

Ind1 <- unique(temp_corr$Ind1)
Ind2 <- Ind1
coefficient <- rep(1.00, times = length(Ind1))
corr_level <- rep("self-comparison", times = length(Ind1))
coeff_label <- rep("1.00",
                   times = length(Ind1))
subgroup <- rep(sub_name,
                   times = length(Ind1))


self_corr <- data.frame(subgroup, Ind1, Ind2, coefficient, corr_level, coeff_label)

heatmap_data <- rbind(temp_corr, self_corr) %>% 
                              mutate(Ind1 = factor(Ind1, 
                                    levels = c("extinct","RLI",
                                                             "threatened",
                                                             "BIIab",
                                                             "BIIri",
                                                             "BHI_plants",
                                                             "HFP"))) %>%
                              mutate(Ind2 = factor(Ind2, levels = c("extinct","RLI",
                                                             "threatened",
                                                             "BIIab",
                                                             "BIIri",
                                                             "BHI_plants",
                                                             "HFP")))

# tri_data <- flipped_correlations %>% 
#       dplyr::select(Ind1, Ind2, coefficient, corr_level, coeff_label) %>% 
#       rbind(self_corr) %>% 
#       mutate(coefficient = ifelse(coefficient == 0.00, "", coefficient))
#  
# tri_data[18,1] <- "HFP"
# tri_data[18,2] <- "extinct"

return(heatmap_data)

}

cluster_data <- split(correlations_clusters, correlations_clusters$subgroup)

clus1 <- prep_heatmap_data(cluster_data[[1]], "Cluster one")
clus2 <- prep_heatmap_data(cluster_data[[2]], "Cluster two")
clus3 <- prep_heatmap_data(cluster_data[[3]], "Cluster three")

cluster_heatmap_data <- rbind(clus1, clus2, clus3) %>% 
                        mutate(subgroup = factor(subgroup, 
                               levels = c("Cluster one","Cluster two",
                                          "Cluster three"))) %>% 
                        mutate(coeff_label = ifelse(coeff_label == "0", "0.00",
                                                    coeff_label)) %>% 
                        mutate(Ind1 = factor(Ind1, 
                               levels = c("BHI_plants","BIIab", "BIIri", "HFP",
                                          "extinct", "RLI", "threatened")),
                               Ind2 = factor(Ind2, 
                               levels = c("BHI_plants","BIIab", "BIIri", "HFP",
                                          "extinct", "RLI", "threatened")))

heatmap_pal <- scales::seq_gradient_pal("gray30", "white", "Lab")(seq(0,1,length.out=4))
heatmap_pal_vir <- c(viridis(4), "grey80")

cluster_heatmap <- ggplot(cluster_heatmap_data, 
                             aes(x = Ind1, y = Ind2, 
                                 fill = corr_level)) +
  geom_tile(color = "white", height = 1.0, width = 1.0, alpha = 0.75) +
  coord_equal() +
  geom_text(aes(label = coeff_label), size = 4) +
  scale_fill_manual(values = heatmap_pal_vir) +
  labs(x = "Indicator 1",
       y = "Indicator 2") +
    guides(fill = guide_legend(title = "Correlation\nstrength")) +
   theme(panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          panel.background = element_rect(fill = "grey97"),
          legend.position = "bottom",
          legend.title=element_text(size=20), 
          legend.text=element_text(size=18),
          axis.title.x = element_text(size=18),
          axis.title.y = element_text(size=18),
          axis.text.x = element_text(size=14, angle = 45, hjust = 1),
          axis.text.y = element_text(size=14),
          axis.ticks.x = element_blank(), 
          axis.ticks.y = element_blank(),
          strip.text.x = element_text(size = 18),
          strip.background = element_blank()) +
          facet_wrap(~ subgroup, nrow = 1)

ggsave(file.path(main_outputs, "cluster_heatmaps.png"),
       cluster_heatmap, width = 16, height = 8, units = "in",
       device = "png")


## Triangle heatmap

prep_heatmap_data_tri <- function(data, sub_name) {
  
inds <- as.data.frame(str_split_fixed(data$combination, " x ", 2)) 

names(inds) <- c("Ind1", "Ind2")

newdata <- cbind(data, inds)  

newdata <- newdata %>% 
           mutate(subgroup = sub_name)

flipped_correlations <- newdata %>% 
                        mutate(Ind12 = Ind1,
                               Ind21 = Ind2) %>% 
                        dplyr::select(-Ind1, -Ind2) %>% 
                        rename(Ind1 = Ind21,
                               Ind2 = Ind12)

temp_corr <- rbind(newdata, flipped_correlations) %>% 
            dplyr::select(subgroup, Ind1, Ind2, coefficient, corr_level, coeff_label)

Ind1 <- unique(temp_corr$Ind1)
Ind2 <- Ind1
coefficient <- rep(1.00, times = length(Ind1))
corr_level <- rep("self-comparison", times = length(Ind1))
coeff_label <- rep("1.00",
                   times = length(Ind1))
subgroup <- rep(sub_name,
                   times = length(Ind1))


self_corr <- data.frame(subgroup, Ind1, Ind2, coefficient, corr_level, coeff_label)

tri_data <- flipped_correlations %>%
      dplyr::select(subgroup,Ind1, Ind2, coefficient, corr_level, coeff_label) %>%
      rbind(self_corr) %>%
      mutate(coefficient = ifelse(coefficient == 0.00, "", coefficient))

return(tri_data)

}

cluster_data <- split(correlations_clusters, correlations_clusters$subgroup)

clus1 <- prep_heatmap_data_tri(cluster_data[[1]], "Cluster one")
clus2 <- prep_heatmap_data_tri(cluster_data[[2]], "Cluster two")
clus3 <- prep_heatmap_data_tri(cluster_data[[3]], "Cluster three")

cluster_heatmap_data <- rbind(clus1, clus2, clus3) %>% 
                        mutate(subgroup = factor(subgroup, 
                               levels = c("Cluster one","Cluster two",
                                          "Cluster three"))) %>% 
                        mutate(coeff_label = ifelse(coeff_label == "0", "0.00",
                                                    coeff_label))%>% 
                        mutate(Ind1 = factor(Ind1, 
                               levels = c("BHI_plants","BIIab", "BIIri", "HFP",
                                          "extinct", "RLI", "threatened")),
                               Ind2 = factor(Ind2, 
                               levels = c("BHI_plants","BIIab", "BIIri", "HFP",
                                          "extinct", "RLI", "threatened")))

pal2 <- c("#357BA2FF", "#49C1ADFF", "#DEF5E5FF", "white","grey80")

cluster_triangles <- ggplot(cluster_heatmap_data, aes(x = Ind1, y = Ind2, 
                                fill = corr_level)) +
  geom_tile(color = "white", height = 1.0, alpha = 0.75) +
  geom_text(aes(label = coefficient), size = 7) +
  scale_fill_manual(values = pal2) +
  labs(x = "Indicator 1",
       y = "Indicator 2") +
    guides(fill = guide_legend(title = "Correlation\nstrength")) +
   theme(panel.grid.major = element_blank(),
         panel.grid.minor = element_blank(),
         panel.background = element_rect(fill = "grey97"),
         legend.position = "bottom",
         legend.title=element_text(size=20), 
         legend.text=element_text(size=18),
         axis.title.x = element_text(size=20),
         axis.title.y = element_text(size=20),
         axis.text.x = element_text(size=18, angle = 45, hjust = 1),
         axis.text.y = element_text(size=18),
         axis.ticks.x = element_blank(), 
         axis.ticks.y = element_blank(),
         strip.text.x = element_text(size = 20))+
         facet_wrap(~ subgroup, nrow = 1)


ggsave(file.path(main_outputs, "cluster_heatmap_triangle.png"),
       cluster_triangles, width = 20, height = 8, units = "in",
       device = "png")


```

# Cluster scatterplots - absolute indicator score

```{r Cluster heatmaps, include = FALSE, echo=FALSE, warning = FALSE, message = FALSE}

correlation_input_data <- indicators_raw_no_lpi %>%
                          merge(ecoregion_clusters[c("ecoregion_id",
                                                     "cluster")],
                                by = "ecoregion_id") %>% 
                          merge(ecoregions_categorical,
                                by = "ecoregion_id") %>% 
                          dplyr::select(-included_in_hfp)

rli_bhi_scatterplot_data <- correlation_input_data %>% 
                            dplyr::select(ecoregion_id,
                                          BIIri,
                                          RLI,
                                          cluster) %>% 
                            mutate(cluster_label = ifelse(
                              cluster == 1, "Cluster one",
                              ifelse(
                              cluster == 2, "Cluster two",
                              ifelse(
                              cluster == 3, "Cluster three",NA)))) %>% 
                            mutate(cluster_label = factor(cluster_label, 
                               levels = c("Cluster one","Cluster two",
                                          "Cluster three")))

cluster_scatterplots <- ggplot(data = rli_bhi_scatterplot_data,
               aes(x = BIIri,
                   y = RLI)) +
        geom_point(shape = 16) +
        stat_cor(method = "spearman", label.x = 0.65, label.y = 0.79) +
        #coord_cartesian(xlim = c(0.3, 1), ylim = c(0.75,1)) +
        theme(panel.grid.major = element_blank(),
              panel.grid.minor = element_blank(),
              panel.background = element_blank(),
              axis.line = element_line(colour = "black"),
              axis.title.x = element_text(size=18),
              axis.title.y = element_text(size=18),
              axis.text.x = element_text(size=14),
              axis.text.y = element_text(size=14),
              strip.text.x = element_text(size = 18),
              strip.background = element_blank()) +
        facet_wrap(~ cluster_label, nrow = 1) +
        #facet_wrap(~ cluster_label, nrow = 1, scales = "free") +
        scale_x_continuous(limits=c(0.3, 1)) + 
        scale_y_continuous(limits=c(0.7,1)) +
        labs(x = "Biodiversity Intactness Index (richness)",
             y = "Red List Index")
  
ggsave(file.path(main_outputs, "cluster_scatterplots.png"),
       cluster_scatterplots, width = 12, height = 6, units = "in",
       device = "png")
```

# Cluster scatterplots - rank order

```{r rank scatterplot, include = FALSE, echo=FALSE, warning = FALSE, message = FALSE}

input_data <- indicators_raw_no_lpi %>%
                          merge(ecoregion_clusters[c("ecoregion_id",
                                                     "cluster")],
                                by = "ecoregion_id") %>% 
                          merge(ecoregions_categorical,
                                by = "ecoregion_id") %>% 
                          dplyr::select(-included_in_hfp)

rank_rli_bhi_scatterplot_data <- input_data %>% 
                            dplyr::select(ecoregion_id,
                                          BIIri,
                                          RLI,
                                          cluster) %>% 
                            mutate(cluster_label = ifelse(
                              cluster == 1, "Cluster one",
                              ifelse(
                              cluster == 2, "Cluster two",
                              ifelse(
                              cluster == 3, "Cluster three",NA)))) %>% 
                            mutate(cluster_label = factor(cluster_label, 
                               levels = c("Cluster one","Cluster two",
                                          "Cluster three"))) %>% 
                            mutate(BHI_rank_order = rank(BIIri, "last"),
                                   RLI_rank_order = rank(RLI))

rank_cluster_scatterplots <- ggplot(data = rank_rli_bhi_scatterplot_data,
               aes(x = BHI_rank_order,
                   y = RLI_rank_order)) +
        geom_point(shape = 16) +
        stat_cor(method = "spearman", label.x = 0.65, label.y = 0.79) +
        coord_cartesian(xlim = c(0, 750), ylim = c(0, 750)) +
        theme(panel.grid.major = element_blank(),
              panel.grid.minor = element_blank(),
              panel.background = element_blank(),
              axis.line = element_line(colour = "black"),
              axis.title.x = element_text(size=18),
              axis.title.y = element_text(size=18),
              axis.text.x = element_text(size=14),
              axis.text.y = element_text(size=14),
              strip.text.x = element_text(size = 18),
              strip.background = element_blank()) +
        facet_wrap(~ cluster_label, nrow = 1)+
        labs(x = "Biodiversity Intactness Index (richness) rank order",
             y = "Red List Index rank order")
  
ggsave(file.path(main_outputs, "rank_cluster_scatterplots.png"),
       rank_cluster_scatterplots, width = 12, height = 6, units = "in",
       device = "png")
```


```{r rank priority ecoregions, echo=FALSE, warning = FALSE, message = FALSE}

top <- -10

bii_priorities <- indicators_raw_no_lpi %>% 
                     merge(ecoregion_clusters[c("ecoregion_id", 
                                                "ECO_NAME",
                                                "cluster")],
                           by = "ecoregion_id") %>% 
                     group_by(cluster) %>% 
                     top_n(top, BIIri) %>% 
                     dplyr::select(ECO_NAME, cluster, BIIri) %>% 
                     ungroup(.) %>% 
                     arrange(cluster)
                      
kable(bii_priorities)

rli_priorities <- indicators_raw_no_lpi %>% 
                     merge(ecoregion_clusters[c("ecoregion_id", 
                                                "ECO_NAME",
                                                "cluster")],
                           by = "ecoregion_id") %>% 
                     group_by(cluster) %>% 
                     top_n(top, RLI) %>% 
                     dplyr::select(ECO_NAME, cluster, RLI) %>% 
                     ungroup(.) %>% 
                     arrange(cluster)
                      
kable(rli_priorities)

```

# Quantiles

```{r quantiles}

# This plot aims to show how many ecoregions are in different quantiles based
# on different indicators (e.g if an ecoregion is in the same quantile for both 
# indicators you might consider there to be consensus, but if they're at extreme
# ends of the quantile it indicates contradiction between indicator scores)

quantile_sf <- indicators_raw_no_lpi %>% 
  merge(ecoregion_clusters[c("ecoregion_id", "cluster", "ECO_NAME")],
                           by = "ecoregion_id") %>% 
  # add year of disturbance
  merge(ecoregions_categorical[c("ecoregion_id","disturbance_year")],
                           by = "ecoregion_id") %>% 
                     group_by(cluster) %>%
  dplyr::select(ecoregion_id, ECO_NAME, cluster, BHI_plants, RLI) %>% 
     mutate(spp_quantile = ntile(RLI, 4),
            lu_quantile = ntile(BHI_plants, 4)) %>% 
   mutate(group = paste(spp_quantile, lu_quantile),
          cluster_n = n()) %>% 
   ungroup(.) %>% 
   group_by(cluster, group) %>% 
   mutate(membership_n = n(),
          prop = membership_n/cluster_n) %>% 
   distinct(.) %>% 
   arrange(desc(cluster)) %>% 
   arrange(desc(prop)) %>% 
   ungroup(.) %>% 
   mutate(agreement_level = lu_quantile - spp_quantile) %>% 
  # Make sf so we can map it later
  merge(ecoregion_geometry,by = "ECO_NAME")

quantile_plot <- ggplot(data = quantile_sf,
               aes(x = lu_quantile,
                   y = spp_quantile)) +
         geom_tile(aes(x = lu_quantile,
                      y = spp_quantile,
                      fill = prop),
                    alpha = 0.8) +
        scale_fill_gradient2(low="blue", mid="white", high="red", 
                       midpoint= 0)+
        geom_jitter(shape = 16, alpha = 0.5) +
        theme(panel.grid.major = element_blank(),
              panel.grid.minor = element_blank(),
              panel.background = element_blank(),
              axis.line = element_line(colour = "black"),
              axis.title.x = element_text(size=18),
              axis.title.y = element_text(size=18),
              axis.text.x = element_text(size=14),
              axis.text.y = element_text(size=14),
              strip.text.x = element_text(size = 18),
              strip.background = element_blank()) +
        facet_wrap(~ cluster, nrow = 1)+
        labs(x = "Land use quantiles",
             y = "Species quantiles")

quantile_plot

ggsave(file.path(supp_outputs, "quantile_plot.png"),
         quantile_plot, device = "png")

quantile_sf <- quantile_sf %>% 
               dplyr::select(ecoregion_id, ECO_NAME, cluster, spp_quantile, 
                             lu_quantile, RLI, BHI_plants, agreement_level, 
                             geometry) %>% 
               st_as_sf()

names(quantile_sf) <- abbreviate(names(quantile_sf), 8)

st_write(quantile_sf, 
         file.path(clean_data, "9_indicator_quantile_rankings.shp"))

disturbed_sf <- quantile_sf %>% 
                dplyr::select(ecorgn_d, ECO_NAME, cluster,geometry) %>% 
                merge(ecoregions_categorical[c("ecoregion_id", 
                                               "disturbance_year")], 
                      by.x = "ecorgn_d",
                      by.y = "ecoregion_id") %>% 
                mutate(disturbed = ifelse(disturbance_year == "undisturbed", 
                                          "undisturbed", "disturbed")) 

# Trying to make polygons of just diturbed/undisturbed locations
# center <- disturbed_sf %>% # select the central parts
#   filter(disturbed == "disturbed") %>% 
#   st_buffer(0.5) %>% # make a buffer of half a meter around all parts (to avoid slivers)
#   st_union() %>% # unite to a geometry object
#   st_sf() %>% # make the geometry a data frame object
#   mutate(disturbed = "disturbed") # return back the data value 
# 
# outskirts <- disturbed_sf%>% # select the central parts
#   filter(disturbed == "undisturbed") %>% 
#   st_buffer(0.5) %>% # make a buffer of half a meter around all parts (to avoid slivers)
#   st_union() %>% # unite to a geometry object
#   st_sf() %>% # make the geometry a data frame object
#   mutate(disturbed = "undisturbed") # return back the data value 
# 
# frmPraha2 <- rbind(center, 
#                    outskirts)
# 
# plot(st_geometry(frmPraha2))
# 
# st_write(frmPraha2,file.path(clean_data, 
#                              "10_disturbed_ecoregions.shp"), append = FALSE)

```

```{r quantile maps}



```

# LPI Scatterplots 

```{r indicator scatterplots, echo=FALSE, warning = FALSE, message = FALSE }

# * Make subgroup scatterplots ----

# Make directories for the scatterplots (this can probably go in the make plots
# loop itself now have resolved issues with naming)

scatter_directories <- list()

for (i in 1:length(names(group_matrices))) {
  
  group <- names(group_matrices)[i]
  
  print(group)
  
  scatter_directory <- file.path(supp_outputs, paste(group,
                                                            "scatterplots", sep = "_"))
  
  if ( !dir.exists( scatter_directory ) ) {
    
    dir.create( scatter_directory, recursive = FALSE )
    
  }
  
  scatter_directories[[i]] <- scatter_directory
  
}

# For each grouping variable, for each subgroup, for each indicator pair,
# make a scatterplot

group_scatterplots <- list()

for (i in seq_along(group_dataframes)){
  
  subgroup_dataframes <- group_dataframes[[i]]
  
  scatter_directory <- scatter_directories[[i]]
  
  subgroup_scatterplots <- list()
  
  for (j in seq_along(subgroup_dataframes)) {
    
    subgroup <- groups[[i]][j]
    
    subgroup
    
    subgroup <- substring(subgroup, 1, 8)
    
    # subgroup <- str_replace(subgroup, " - ", "_")
    # subgroup <- str_replace(subgroup, " & ", "_")
    # subgroup <- str_replace(subgroup, " , ", "_")
    # subgroup <- str_replace(subgroup, " . ", "_")
    # subgroup <- str_replace(subgroup, " ", "_")
    # subgroup <- str_replace(subgroup, ", ", "_")
    # subgroup <- str_replace(subgroup, " ", "_")
    # 
    # subgroup
    
    subgroup_scatterplots[[j]] <- make_subgroup_scatterplot(subgroup_dataframes[[j]], 
                                                            names(group_dataframes)[i], 
                                                            subgroup,
                                                            scatter_directory,
                                                            "lpi")
    
  }
  
  group_scatterplots[[i]] <- subgroup_scatterplots
  
}

```
