==> 6_maps_and_methods.Rmd <==
---
title: "Maps and methods"
author: "Simone Stevenson"
date: "04/05/2021"
output: html_document
---

```{r setup, include=FALSE}

# Load packages

## Data wrangling
library(tidyverse)
library(tidylog)

## Data visualisation
library(tmap)
library(viridisLite)

# Set up inputs

date <- Sys.Date()
today <- date
location <- "global"
country <- NA #"Australia" # If not subsetting, set as NA, e.g. country <- NA
#analysis_inputs <- "N:/Quantitative-Ecology/Simone/extinction_test/outputs/version_3/2020-08-25_indicator_output_files"
analysis_inputs <- "N:/LES/Burwood/Quantitative-Ecology/Simone/extinction_test/outputs/version_3/2020-08-25_indicator_output_files"
save_outputs <- "yes" #only applies to maps, other things will always save
#outputs <- "N:/Quantitative-Ecology/Simone/extinction_test/outputs"
outputs <- "N:/LES/Burwood/Quantitative-Ecology/Simone/extinction_test/outputs"
timepoints <- c("2005", "2008")
#manuscript_outputs <- "N:\\Quantitative-Ecology\\Simone\\extinction_test\\outputs\\version_3\\2021-05-04_manuscript_outputs"
manuscript_outputs <- "N:/LES/Burwood/Quantitative-Ecology\\Simone\\extinction_test\\outputs\\version_3\\2021-05-04_manuscript_outputs"
  
dir.create(manuscript_outputs, recursive = TRUE ) # create a new directory for today's outputs
  
  
main_outputs <- file.path(manuscript_outputs, paste(date,
                                 "_main_manuscript_outputs",sep="") )
  
dir.create(main_outputs, recursive = TRUE )

supp_outputs <- file.path(manuscript_outputs, paste(date,
                                 "_supporting_info_outputs",sep="") )
  
dir.create(supp_outputs, recursive = TRUE )



indicator_values_master <- readRDS(file.path(analysis_inputs,
                          "global_ecoregions_2017_indicator_values_master.rds"))

ecoregion_values_master <- readRDS(file.path(analysis_inputs,
                               "global_ecoregions_2017_ecoregion_values_master.rds"))

ecoregion_map_all <- readRDS(paste(file.path("N:/LES/Burwood/Quantitative-Ecology/Simone/extinction_test/inputs",  "ecoregions_2017"),"Ecoregions2017valid.rds"))

ecoregion_map_data <- ecoregion_map_all %>%
                 dplyr::select(ECO_ID, ECO_NAME, OBJECTID, REALM, geometry) %>%
                      rename(ecoregion_id = ECO_ID) %>%
                      dplyr::select(ecoregion_id, geometry)

rm(ecoregion_map_all)


indicator_values_map <- indicator_values_master %>%
                        dplyr::select(ecoregion_id, indicator, year, raw_indicator_value) %>%
                        distinct(.)

indicator_map_data <- left_join(ecoregion_map_data, indicator_values_map,
                                by = "ecoregion_id")

names(indicator_map_data)

# ecoregion_map_data <- ecoregion_map %>%
#   rename(ecoregion_id = ECO_ID) %>%
#   dplyr::select(ecoregion_id, geometry)

ecoregion_values_map <- ecoregion_values_master %>%
  dplyr::select(ecoregion_id, indicator, year, raw_indicator_value) %>%
  distinct(.)

eco_variable_map_data <- left_join(ecoregion_map_data, ecoregion_values_map,
                                by = "ecoregion_id")


```

## Biodiversity Intactness Index - Richness

```{r}
library(ggplot2)
library(viridis)
library(sf)
library(ggspatial)

# Function to create a world map with graticules
create_world_map <- function(sf_object, bg, label, d, pal, keytitle) {
  
    ggplot() +
    geom_sf(data = sf_object, aes(fill = raw_indicator_value,
                                  col = raw_indicator_value)) +
    scale_fill_viridis(direction = d, option = pal, na.value = "gray",
                       breaks = pretty(sf_object$raw_indicator_value, n = 5)) +  # Reverse the color scale
    scale_color_viridis(direction = d, option = pal, na.value = "gray",
                        breaks = pretty(sf_object$raw_indicator_value, n = 5),
                        guide = "none") +  # Reverse the color scale
    labs(title = label,
         fill = keytitle) +
    theme(legend.text = element_text(size = 10), 
          plot.title = element_text(size = 16, hjust = 0.5),
          plot.subtitle = element_text(size = 12, hjust = 0.5),
          panel.background = element_rect(fill = bg, color = "black"),
          plot.background = element_rect(fill = "white")) +  # Set panel background color
    # annotation_scale(location = "br", width_hint = 0.1) +  # Add scale bar
    # annotation_north_arrow(location = "br", which_north = "true", 
    #                        height = unit(5, "mm"),
    #                        width = unit(5, "mm"),
    #                        pad_y = unit(10, "mm")) +  # Add north arrow
    theme(legend.background = element_rect(fill = "white"),
          panel.grid = element_blank()) +
    guides(fill = guide_legend(override.aes = list(size = 1, ncol = 5))) +
    theme(legend.position = "bottom") 
}


# Call the function with your sf object
# create_world_map(test_data, "test", "seashell")
```

```{r}
# Function to create a world map with graticules
create_world_map_categorical <- function(sf_object, bg, label, d, pal, keytitle) {
  
    ggplot() +
    geom_sf(data = sf_object, aes(fill = raw_indicator_value,
                                  col = raw_indicator_value)) +
    scale_fill_viridis_d(direction = d, option = pal, na.value = "gray",
                       breaks = pretty(sf_object$raw_indicator_value, n = 5)) +  # Reverse the color scale
    scale_color_viridis_d(direction = d, option = pal, na.value = "gray",
                        breaks = pretty(sf_object$raw_indicator_value, n = 5),
                        guide = "none") +  # Reverse the color scale
    labs(title = label,
         fill = keytitle) +
    theme(legend.text = element_text(size = 10), 
          plot.title = element_text(size = 16, hjust = 0.5),
          plot.subtitle = element_text(size = 12, hjust = 0.5),
          panel.background = element_rect(fill = bg, color = "black"),
          plot.background = element_rect(fill = "white")) +  # Set panel background color
    # annotation_scale(location = "br", width_hint = 0.1) +  # Add scale bar
    # annotation_north_arrow(location = "br", which_north = "true", 
    #                        height = unit(5, "mm"),
    #                        width = unit(5, "mm"),
    #                        pad_y = unit(10, "mm")) +  # Add north arrow
    theme(legend.background = element_rect(fill = "white"),
          panel.grid = element_blank()) +
    guides(fill = guide_legend(override.aes = list(size = 1, ncol = 5))) +
    theme(legend.position = "bottom") 
}


```

```{r BII richness, echo = FALSE}

# * BII Richness ----

    rbii_data <- indicator_map_data %>%
      filter(indicator == "richness biodiversity intactness index" & year == 2005)

# simplified data for testing

rbii <- create_world_map(rbii_data, "snow", "BII (richness)", 1, "viridis", "Indicator value")

ggsave(file.path(supp_outputs, paste(today, location,
                               "richness_BII_map.png", sep = "_")),
       rbii, width=14, height=10, unit = "cm")
    
    # rbii <- tm_shape(rbii_data) +
    #   tm_polygons(col = "raw_indicator_value",
    #               border.col = "black",
    #               style = "cont",
    #               pal = "viridis",
    #               title = "Biodiversity\nIntactness Index\n(Richness)") +
    #   tm_layout(legend.outside = TRUE,
    #             legend.outside.position = "right") 
    # 
    # rbii
    # 
    # tmap_save(rbii, file.path(supp_outputs, paste(today, location,
    #                            "richness_BII_map.png", sep = "_")),
    #           width=2000, height=1080, asp=0)
    # 
    # rm(rbii, rbii_data)


```

## Biodiversity Intactness Index - Abundance

```{r BII abundance, echo = FALSE}
# * BII Abundance ----

abii_data <- indicator_map_data %>%
      filter(indicator == "abundance biodiversity intactness index" & year == 2005)

abii <- create_world_map(rbii_data, "snow", "BII (abundance)", 1, "viridis")

ggsave(file.path(supp_outputs, paste(today, location,
                               "abundance_BII_map.png", sep = "_")),
       abii, width=14, height=10, unit = "cm")
    
    # abii <- tm_shape(abii_data) +
    #   tm_polygons(col = "raw_indicator_value",
    #               border.col = "black",
    #               style = "cont",
    #               pal = "viridis",
    #               title = "Biodiversity\nIntactness Index\n(Abundance)") +
    #   tm_layout(legend.outside = TRUE,
    #             legend.outside.position = "right") +
    #   tm_borders(lwd = 0.5)
    # 
    # abii
    # 
    # tmap_save(abii, file.path(supp_outputs, paste(today, location,
    #                           "abundance_BII_map.png", sep = "_")),
    #           width=2000, height=1080, asp=0)
    
rm(abii_data, abii)
    ```
    
## BHI Plants 

```{r BHI, echo = FALSE}

bhi_data <- indicator_map_data %>%
      filter(indicator == "BHI plants" & year == 2005)

bhi <- create_world_map(bhi_data, "snow", "BHI (plants)", 1, "viridis")

ggsave(file.path(supp_outputs, paste(today, location,
                               "BHI_map.png", sep = "_")),
       bhi, width=14, height=10, unit = "cm")
    
    # bhi <- tm_shape(bhi_data) +
    #   tm_polygons(col = "raw_indicator_value",
    #               border.col = "black",
    #               style = "cont",
    #               pal = "viridis",
    #               title = "Biodiversity\nHabitat Index\n(Plants)") +
    #   tm_layout(legend.outside = TRUE,
    #             legend.outside.position = "right") 
    # 
    # bhi
    # 
    # tmap_save(bhi, file.path(indicator_outputs, paste(location,
    #                         "BHI_plants_map.png", sep = "_")),
    #           width=1920, height=1080, asp=0)

rm(bhi, bhi_data)

```

# Proportion threatened

```{r proportion threatened, echo = FALSE}
threatened_data <- indicator_map_data %>%
              filter(indicator == "proportion at risk" & year == 2008)

threatened <- create_world_map(threatened_data, "snow", "Proportion threatened", -1, "viridis")

ggsave(file.path(supp_outputs, paste(today, location,
                               "threatened_map.png", sep = "_")),
       threatened, width=14, height=10, unit = "cm")

# threatened <- tm_shape(threatened_data) +
#               tm_polygons(col = "raw_indicator_value",
#                            border.col = "black",
#                            style = "order",
#                            pal = "-viridis",
#                            title = "Proportion of\nspecies threatened ") +
#               tm_layout(legend.outside = TRUE,
#                          legend.outside.position = "right") 
# 
# threatened

# tmap_save(threatened, file.path(indicator_outputs, paste(location,
#                        "proportion_threatened_map.png", sep = "_")),
#                        width=1920, height=1080, asp=0)

rm(threatened, threatened_data)
```

## Proportion extinct 

```{r proportion extinct, echo = FALSE}

extinct_data <- indicator_map_data %>%
  filter(indicator == "proportion extinct" & year == 2008) 

extinct <- create_world_map(extinct_data, "snow", "Proportion extinct", -1, "viridis")

ggsave(file.path(supp_outputs, paste(today, location,
                               "extinct_map.png", sep = "_")),
       extinct, width=14, height=10, unit = "cm")



# 
# extinct <- tm_shape(extinct_data) +
#            tm_polygons(col = "raw_indicator_value",
#                        border.col = "black",
#                        style = "fixed",
#                        breaks = c(0,0,0.01, 0.02, 0.03, 0.04, 
#                                   0.05, 0.06, 0.07, 0.08, 0.09, 0.1),
#                        pal = "-viridis",
#                        title = "Proportion of\nspecies extinct",
#                        alpha = 0.8,
#                        lwd = 0.1, 
#                        colorNA = "grey") +
#            tm_layout(legend.outside = TRUE,
#                      legend.outside.position = "right") 
# 
# extinct
# 
# tmap_save(extinct, file.path(indicator_outputs, paste(location,
#                    "proportion_extinct_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)

rm(extinct_data, extinct)
```

## Number extinct

```{r number extinct, echo = FALSE}
# * Number extinct ----

num_extinct_data <- indicator_map_data %>%
  filter(indicator == "number extinct" & year == 2008)

num_extinct <- create_world_map(num_extinct_data, "snow", "Number extinct", -1, "viridis")

ggsave(file.path(supp_outputs, paste(today, location,
                               "number_extinct_map.png", sep = "_")),
       num_extinct, width=14, height=10, unit = "cm")


# num_extinct <- tm_shape(num_extinct_data) +
#   tm_polygons(col = "raw_indicator_value",
#               border.col = "black",
#               style = "order",
#               pal = "-viridis",
#               title = "Number of\nspecies extinct") +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right") 
# 
# num_extinct
# 
# tmap_save(num_extinct, file.path(indicator_outputs, paste(location,
#           "number_extinct_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)

rm(num_extinct_data, num_extinct)

```

# Red List Index (all taxa)

```{r RLI all, echo = FALSE}
# * RLI ----

# ** All groups ----

rli_data <- indicator_map_data %>%
  filter(indicator == "RLI" & year == 2008)

rli <- create_world_map(rli_data, "snow", "RLI", 1, "viridis") 

ggsave(file.path(supp_outputs, paste(today, location,
                               "RLI_map.png", sep = "_")),
       rli, width=14, height=10, unit = "cm")

# rli <- tm_shape(rli_data) +
#   tm_polygons(col = "raw_indicator_value",
#               border.col = "black",
#               style = "order",
#               pal = "viridis",
#               title = "Red List Index") +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right") 
# 
# rli
# 
# tmap_save(rli, file.path(indicator_outputs, paste(location,
#                "RLI_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)

rm(rli_data, rli)
```

## Red List Index Birds

```{r RLI birds, echo = FALSE}
# bird_rli_data <- indicator_map_data %>%
#   filter(indicator == "RLI-birds" & year == 2008)
# 
# bird_rli <- tm_shape(bird_rli_data) +
#   tm_polygons(col = "raw_indicator_value",
#               border.col = "black",
#               style = "order",
#               pal = "viridis",
#               title = "Bird Red List Index") +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right") 
# 
# bird_rli
# 
# tmap_save(bird_rli, file.path(indicator_outputs, paste(location,
#                                                   "bird_RLI_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)
# 
# rm(bird_rli_data, bird_rli)
```

## Red List Index Mammals

```{r BII abundance, echo = FALSE}
# ** Mammal RLI ----

# mammal_rli_data <- indicator_map_data %>%
#   filter(indicator == "RLI-mammals" & year == 2008)
# 
# mammal_rli <- tm_shape(mammal_rli_data) +
#   tm_polygons(col = "raw_indicator_value",
#               border.col = "black",
#               style = "order",
#               pal = "viridis",
#               title = "Mammal Red List Index") +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right") 
# 
# mammal_rli
# 
# tmap_save(mammal_rli, file.path(indicator_outputs, paste(location,
#                                                        "mammal_RLI_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)
# 
# rm(mammal_rli_data, mammal_rli)

```

## Red List Index Amphibians

```{r rli amphibians, echo = FALSE}
# ** Amphibian RLI ----

# amphibian_rli_data <- indicator_map_data %>%
#   filter(indicator == "RLI-amphibians" & year == 2008)
# 
# amphibian_rli <- tm_shape(amphibian_rli_data) +
#   tm_polygons(col = "raw_indicator_value",
#               border.col = "black",
#               style = "order",
#               pal = "viridis",
#               title = "Amphibian Red List Index") +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right") 
# 
# amphibian_rli
# 
# tmap_save(amphibian_rli, file.path(indicator_outputs, paste(location,
#                                                        "amphibian_RLI_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)
# 
# rm(amphibian_rli_data, amphibian_rli)

```

## Human Footprint Index

```{r hfp, echo = FALSE}
# * HFP ----

hfp_data <- indicator_map_data %>%
  filter(indicator == "mean human footprint index" & year == 2005)

hfp <- create_world_map(hfp_data, "snow", "HFP", -1, "viridis")

ggsave(file.path(supp_outputs, paste(today, location,
                               "human_footprint_map.png", sep = "_")),
       hfp, width=14, height=10, unit = "cm")

# hfp <- tm_shape(hfp_data) +
#   tm_polygons(col = "raw_indicator_value",
#               border.col = "black",
#               style = "order",
#               pal = "-viridis",
#               title = "Human\nFootprint Index") +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right") 
# 
# hfp
# 
# tmap_save(hfp, file.path(indicator_outputs, paste(location,
#                          "human_footprint_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)

rm(hfp_data, hfp)
```

## Living Planet Index

```{r LPI, echo = FALSE}
# * LPI ----

lpi_map_data <- indicator_map_data %>%
  filter(indicator == "LPI" & year == 2005) %>%
  st_drop_geometry()

lpi_map_data <- ecoregion_map_data[c("ecoregion_id", "geometry")] %>%
            merge(lpi_map_data, by.x = "ecoregion_id", by.y = "ecoregion_id" ,
                  all.x = TRUE) %>%
            mutate(indicator = "LPI")

lpi <- create_world_map(lpi_map_data, "snow", "LPI", 1, "viridis")

ggsave(file.path(supp_outputs, paste(today, location,
                               "LPI_map.png", sep = "_")),
       lpi, width=14, height=10, unit = "cm")

# head(lpi_map_data)
# 
# lpi <- tm_shape(lpi_map_data) +
#        tm_polygons(col = "raw_indicator_value",
#                     border.col = "black",
#                     style = "order",
#                     pal = "viridis",
#                     title = "Living Planet Index") +
#        tm_layout(legend.outside = TRUE,
#                   legend.outside.position = "right")
# 
# lpi
# 
# tmap_save(lpi, file.path(indicator_outputs, paste(location,
#                           "LPI_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)
# 
# rm(lpi_data, lpi)
```

# Ecoregion variables

## Islands
```{r islands, echo = FALSE}

# * Islands ----

island_data <- eco_variable_map_data %>%
  filter(indicator == "island status")

ipl <- viridis_pal(option = "plasma")(length(unique(island_data$raw_indicator_value)))


islands <-  ggplot() +
    geom_sf(data = island_data, aes(fill = raw_indicator_value,
                                  col = raw_indicator_value)) +
    scale_fill_manual(values = ipl)+ 
    scale_color_manual(values = ipl, guide = "none")+
    labs(title = "Island status",
         fill = "Ecoregion Value") +
    theme(legend.text = element_text(size = 10), 
          plot.title = element_text(size = 16, hjust = 0.5),
          plot.subtitle = element_text(size = 12, hjust = 0.5),
          panel.background = element_rect(fill = "snow", color = "black"),
          plot.background = element_rect(fill = "white")) +  # Set panel background color
    theme(legend.background = element_rect(fill = "white"),
          panel.grid = element_blank()) +
    #guides(fill = guide_legend(override.aes = list(size = 1, ncol = 5))) +
    theme(legend.position = "right")

ggsave(file.path(supp_outputs, paste(today, location,
                               "island_map.png", sep = "_")),
       islands, width=14, height=10, unit = "cm")

# 
# island <- tm_shape(island_data) +
#   tm_polygons(col = "raw_indicator_value",
#               border.col = "black",
#               pal = "-viridis",
#               title = "Island status",
#               colorNA = "grey",
#               alpha = 0.8,
#               lwd = 0.05) +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right",
#             bg.color = "lightblue1") 
# 
# island
# 
# tmap_save(island, file.path(indicator_outputs, paste(location,
#                                                     "island_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)

rm(island, island_data)
```

## Biomes

```{r biomes, echo = FALSE}
biome_data <- eco_variable_map_data %>%
  filter(indicator == "Biome")

bpl <- viridis_pal(option = "plasma")(length(unique(biome_data$raw_indicator_value)))

biomes <-  ggplot() +
    geom_sf(data = biome_data, aes(fill = raw_indicator_value,
                                  col = raw_indicator_value)) +
    scale_fill_manual(values = bpl,guide = guide_legend(ncol = 2, 
                                                        override.aes = list(size = 4)))+ 
    scale_color_manual(values = bpl, guide = "none")+
    labs(title = "Biomes",
         fill = "Ecoregion Value") +
    theme(legend.text = element_text(size = 4), 
          plot.title = element_text(size = 16, hjust = 0.5),
          plot.subtitle = element_text(size = 12, hjust = 0.5),
          panel.background = element_rect(fill = "snow", color = "black"),
          plot.background = element_rect(fill = "white"),
          panel.grid = element_blank()) +  # Set panel background color
    theme(legend.background = element_rect(fill = "white")) +
    #guides(fill = guide_legend(override.aes = list(size = 1, ncol = 5))) +
    theme(legend.position = "bottom") + theme(legend.key.size = unit(0.15, "cm"))

ggsave(file.path(supp_outputs, paste(today, location,
                               "biome_map.png", sep = "_")),
       biomes, width=14, height=10, unit = "cm")

# biome <- tm_shape(biome_data) +
#   tm_polygons(col = "raw_indicator_value",
#               border.col = "black",
#               pal = "viridis",
#               title = "Biome",
#               colorNA = "grey",
#               alpha = 0.8) +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right") 
# 
# biome
# 
# tmap_save(biome, file.path(indicator_outputs, paste(location,
#           "biome_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)

rm(biome, biome_data)
```

## Realms

```{r realms, echo = FALSE}

realm_data <- eco_variable_map_data %>%
  filter(indicator == "realm")

rpl <- viridis_pal(option = "plasma")(length(unique(realm_data$raw_indicator_value)))

realms <-  ggplot() +
    geom_sf(data = realm_data, aes(fill = raw_indicator_value,
                                  col = raw_indicator_value)) +
  scale_fill_manual(values = rpl)+ 
  scale_color_manual(values = rpl, guide = "none")+
  labs(title = "Realms",
         fill = "Ecoregion Value") +
    theme(legend.text = element_text(size = 10), 
          plot.title = element_text(size = 16, hjust = 0.5),
          plot.subtitle = element_text(size = 12, hjust = 0.5),
          panel.background = element_rect(fill = "snow", color = "black"),
          plot.background = element_rect(fill = "white")) +  # Set panel background color
    theme(legend.background = element_rect(fill = "white"),
          panel.grid = element_blank()) +
    #guides(fill = guide_legend(override.aes = list(size = 1, ncol = 5))) +
    theme(legend.position = "right") 

ggsave(file.path(supp_outputs, paste(today, location,
                               "realm_map.png", sep = "_")),
       realms, width=14, height=10, unit = "cm")

# realm <- tm_shape(realm_data) +
#   tm_polygons(col = "raw_indicator_value",
#               style = "cont",
#               border.col = "black",
#               pal = "viridis",
#               title = "Realm",
#               colorNA = "grey",
#               alpha = 0.8) +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right") 
# 
# realm
# 
# tmap_save(realm, file.path(indicator_outputs, paste(location,
#           "realm_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)

rm(realm, realm_data)

```

## Number of RLI records

```{r BII abundance, echo = FALSE}
rli_record_data <- eco_variable_map_data %>%
  filter(indicator == "RLI_records") %>%
  mutate(raw_indicator_value = as.numeric(raw_indicator_value))


rli_record <- create_world_map(rli_record_data, "snow", "RLI # records", 1, "plasma", "Ecoregion value")

ggsave(file.path(supp_outputs, paste(today, location,
                               "rli_record_map.png", sep = "_")),
       rli_record, width=14, height=10, unit = "cm")

# rli_record <- tm_shape(rli_record_data) +
#   tm_polygons(col = "raw_indicator_value",
#               style = "order",
#               border.col = "black",
#               pal = "viridis",
#               title = "RLI records",
#               colorNA = "grey",
#               alpha = 0.8) +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right") 
# 
# rli_record
# 
# tmap_save(rli_record, file.path(indicator_outputs, paste(location,
#           "rli_record_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)

rm(rli_record, rli_record_data)
```

## Predominant threat type

```{r BII abundance, echo = FALSE}
threat_data <- eco_variable_map_data %>%
  filter(indicator == "predominant threat type")

thpl <- viridis_pal(option = "plasma")(length(unique(threat_data$raw_indicator_value)))

threat <- ggplot() +
    geom_sf(data = threat_data, aes(fill = raw_indicator_value,
                                  col = raw_indicator_value)) +
 scale_fill_manual(values = thpl,guide = guide_legend(ncol = 2, 
                                                        override.aes = list(size = 5,
                                                                            key.size = 0.05)))+ 
    scale_color_manual(values = thpl, guide = "none")+
    labs(title = "Predominant threat type",
         fill = "Ecoregion Value") +
    theme(legend.text = element_text(size = 4), 
          plot.title = element_text(size = 16, hjust = 0.5),
          plot.subtitle = element_text(size = 12, hjust = 0.5),
          panel.background = element_rect(fill = "snow", color = "black"),
          plot.background = element_rect(fill = "white")) +  # Set panel background color
    theme(legend.background = element_rect(fill = "white"),
          panel.grid = element_blank()) +
    #guides(fill = guide_legend(override.aes = list(size = 1, ncol = 5))) +
    theme(legend.position = "bottom") 

ggsave(file.path(supp_outputs, paste(today, location,
                               "predominant_threat_map.png", sep = "_")),
       threat, width=14, height=10, unit = "cm")

# threat <- tm_shape(threat_data) +
#   tm_polygons(col = "raw_indicator_value",
#               border.col = "black",
#               pal = "viridis",
#               title = "Predominant threat type",
#               colorNA = "grey",
#               alpha = 0.8) +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "bottom") 
# 
# threat
# 
# tmap_save(threat, file.path(indicator_outputs, paste(location,
#                   "predominant_threat_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)

rm(threat, threat_data)
```

## Headline threat type

```{r BII abundance, echo = FALSE}
headline_threat_data <- eco_variable_map_data %>%
  filter(indicator == "headline threat type")

hthpl <- viridis_pal(option = "plasma")(length(unique(headline_threat_data$raw_indicator_value)))

head_threat <- ggplot() +
    geom_sf(data = headline_threat_data, aes(fill = raw_indicator_value,
                                  col = raw_indicator_value)) +
 scale_fill_manual(values = hthpl,guide = guide_legend(ncol = 2, 
                                                        override.aes = list(size = 5,
                                                          key.size = 0.05)))+ 
    scale_color_manual(values = hthpl, guide = "none")+
    labs(title = "Headline threat type",
         fill = "Ecoregion Value") +
    theme(legend.text = element_text(size = 4), 
          plot.title = element_text(size = 16, hjust = 0.5),
          plot.subtitle = element_text(size = 12, hjust = 0.5),
          panel.background = element_rect(fill = "snow", color = "black"),
          plot.background = element_rect(fill = "white")) +  # Set panel background color
    theme(legend.background = element_rect(fill = "white"),
          panel.grid = element_blank()) +
    #guides(fill = guide_legend(override.aes = list(size = 1, ncol = 5))) +
    theme(legend.position = "bottom") 
ggsave(file.path(supp_outputs, paste(today, location,
                               "headline_threat_map.png", sep = "_")),
       head_threat, width=14, height=10, unit = "cm")

# headline_threat <- tm_shape(headline_threat_data) +
#   tm_polygons(col = "raw_indicator_value",
#               border.col = "black",
#               pal = "viridis",
#               title = "Headline threat type",
#               colorNA = "grey",
#               alpha = 0.8) +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right") 
# 
# headline_threat
# 
# tmap_save(headline_threat, file.path(indicator_outputs, paste(location,
#            "headline_threat_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)

rm(head_threat, headline_threat_data)

```

## Included in HFP

```{r included in hfp, echo = FALSE}
incl_hfp_data <- eco_variable_map_data %>%
  filter(indicator == "included in HFP")

ihpl <- viridis_pal(option = "plasma")(length(unique(incl_hfp_data$raw_indicator_value)))

incl_hfp <- ggplot() +
    geom_sf(data = incl_hfp_data, aes(fill = raw_indicator_value,
                                  col = raw_indicator_value)) +
 scale_fill_manual(values = ihpl,guide = guide_legend(ncol = 2, 
                                                        override.aes = list(size = 5,
                                                                            key.size = 0.05)))+ 
    scale_color_manual(values = ihpl, guide = "none")+
    labs(title = "Included in HFP",
         fill = "Ecoregion Value") +
    theme(legend.text = element_text(size = 4), 
          plot.title = element_text(size = 16, hjust = 0.5),
          plot.subtitle = element_text(size = 12, hjust = 0.5),
          panel.background = element_rect(fill = "snow", color = "black"),
          plot.background = element_rect(fill = "white")) +  # Set panel background color
    theme(legend.background = element_rect(fill = "white"),
          panel.grid = element_blank()) +
    #guides(fill = guide_legend(override.aes = list(size = 1, ncol = 5))) +
    theme(legend.position = "bottom")

ggsave(file.path(supp_outputs, paste(today, location,
                               "hfp_inclusion_map.png", sep = "_")),
       incl_hfp, width=14, height=10, unit = "cm")

# 
# incl_hfp <- tm_shape(incl_hfp_data) +
#   tm_polygons(col = "raw_indicator_value",
#               border.col = "black",
#               pal = "viridis",
#               title = "Included in HFP",
#               colorNA = "grey",
#               alpha = 0.8) +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right") 
# 
# incl_hfp
# 
# tmap_save(incl_hfp, file.path(indicator_outputs, paste(location,
#           "hfp_inclusion_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)
# 
# rm(incl_hfp, incl_hfp_data)

```

## Number of endemics

```{r BII abundance, echo = FALSE}
# endemics_data <- eco_variable_map_data %>%
#   filter(indicator == "number of endemics")%>%
#   mutate(raw_indicator_value = as.numeric(raw_indicator_value))
# 
# endemics <- tm_shape(endemics_data) +
#   tm_polygons(col = "raw_indicator_value",
#               border.col = "black",
#               style = "fixed",
#               breaks = c(0, 0, 1, 5,10, 15, 20, 25, 30, 80),
#               pal = "viridis",
#               title = "Number of endemics",
#               colorNA = "grey",
#               alpha = 0.8) +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right") 
# 
# endemics
# 
# tmap_save(endemics, file.path(indicator_outputs, paste(location,
#           "endemics_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)
# 
# rm(endemics, endemics_data)

```

## Mean human population density

```{r human, echo = FALSE}

# human_pop_data <- eco_variable_map_data %>%
#   filter(indicator == "mean human population density")%>%
#   mutate(raw_indicator_value = as.numeric(raw_indicator_value))
# 
# human_pop <- tm_shape(human_pop_data) +
#   tm_polygons(col = "raw_indicator_value",
#               style = "log10_pretty",
#               border.col = "black",
#               pal = "viridis",
#               title = "Mean human population density",
#               colorNA = "grey",
#               alpha = 0.8) +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right") 
# 
# human_pop
# 
# tmap_save(human_pop, file.path(indicator_outputs, paste(location,
#           "human_population_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)
# 
# rm(human_pop, human_pop_data)

```

## Number of LPI records

```{r lpi records, echo = FALSE}

lpi_records_data <- eco_variable_map_data %>%
  filter(indicator == "LPI_records") %>%
  mutate(raw_indicator_value = as.numeric(raw_indicator_value))

lpi_records <- create_world_map(lpi_records_data, "snow", "LPI # populations",
                                1, "plasma", "Ecoregion value")

ggsave(file.path(supp_outputs, paste(today, location,
                               "lpi_records_map.png", sep = "_")),
       lpi_records, width=14, height=10, unit = "cm")

# lpi_records <- tm_shape(lpi_records_data) +
#   tm_polygons(col = "raw_indicator_value",
#               style = "fixed",
#               breaks = c(0, 0, 25, 50, 100, 200, 300, 400),
#               border.col = "black",
#               pal = "viridis",
#               title = "LPI records",
#               colorNA = "grey",
#               alpha = 0.8) +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right") 
# 
# lpi_records
# 
# tmap_save(lpi_records, file.path(indicator_outputs, paste(location,
#           "lpi_records_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)

rm(lpi_records, lpi_records_data)
```

# Disturbance year

```{r disturbance year, echo = FALSE}

anthrome_data <- eco_variable_map_data %>%
  filter(indicator == "disturbance year")

apl <- viridis_pal(option = "plasma")(length(unique(anthrome_data$raw_indicator_value)))

anthrome <- ggplot() +
    geom_sf(data = anthrome_data, aes(fill = raw_indicator_value,
                                  col = raw_indicator_value)) +
  scale_fill_manual(values = apl)+ 
  scale_color_manual(values = apl, guide = "none")+
  labs(title = "Urbanisation year",
         fill = "Ecoregion Value") +
    theme(legend.text = element_text(size = 10), 
          plot.title = element_text(size = 16, hjust = 0.5),
          plot.subtitle = element_text(size = 12, hjust = 0.5),
          panel.background = element_rect(fill = "snow", color = "black"),
          plot.background = element_rect(fill = "white")) +  # Set panel background color
    theme(legend.background = element_rect(fill = "white"),
          panel.grid = element_blank()) +
    #guides(fill = guide_legend(override.aes = list(size = 1, ncol = 5))) +
    theme(legend.position = "bottom") 


ggsave(file.path(supp_outputs, paste(today, location,
                               "disturbance_year_map.png", sep = "_")),
       anthrome, width=14, height=10, unit = "cm")

# anthrome <- tm_shape(anthrome_data) +
#   tm_polygons(col = "raw_indicator_value",
#               border.col = "black",
#               pal = "viridis",
#               title = "Disturbance year",
#               colorNA = "grey",
#               alpha = 0.8) +
#   tm_layout(legend.outside = TRUE,
#             legend.outside.position = "right") 
# 
# anthrome
# 
# tmap_save(anthrome, file.path(indicator_outputs, paste(location,
#                               "disturbance_year_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)

rm(anthrome, anthrome_data)

```

## High beta diversity 

```{r beta, echo = FALSE}
# beta_data <- eco_variable_map_data %>%
#   filter(indicator == "High beta area")
# 
# beta_data$raw_indicator_value <- as.numeric(beta_data$raw_indicator_value)
# 
# beta <- tm_shape(beta_data) +
#         tm_polygons(col = "raw_indicator_value",
#                     border.col = "black",
#                     style = "cont",
#                     pal = "viridis",
#                     title = "High beta area km^2",
#                     colorNA = "grey",
#                     alpha = 0.8) +
#         tm_layout(legend.outside = TRUE,
#                   legend.outside.position = "right") 
# 
# beta
# 
# 
# tmap_save(beta, file.path(indicator_outputs, paste(location,
#                               "beta_diversity_map.png", sep = "_")),
#           width=1920, height=1080, asp=0)
```
